% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Account.childLink}
\alias{Account.childLink}
\title{Account.childLink Object}
\usage{
Account.childLink(href = NULL, type = NULL)
}
\arguments{
\item{href}{Link to the list of web properties for this account}

\item{type}{Type of the child link}
}
\value{
Account.childLink object
}
\description{
Account.childLink Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Child link for an account entry. Points to the list of web properties for this account.
}
\seealso{
Other Account functions: \code{\link{Account.permissions}},
  \code{\link{Account}}
}
