% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Account.permissions}
\alias{Account.permissions}
\title{Account.permissions Object}
\usage{
Account.permissions()
}
\value{
Account.permissions object
}
\description{
Account.permissions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Permissions the user has for this account.
}
\seealso{
Other Account functions: \code{\link{Account.childLink}},
  \code{\link{Account}}
}

