% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Account}
\alias{Account}
\title{Google Analytics API Objects 
Views and manages your Google Analytics data.}
\usage{
Account(Account.childLink = NULL, Account.permissions = NULL,
  childLink = NULL, created = NULL, id = NULL, name = NULL,
  permissions = NULL, selfLink = NULL, starred = NULL, updated = NULL)
}
\arguments{
\item{Account.childLink}{The \link{Account.childLink} object or list of objects}

\item{Account.permissions}{The \link{Account.permissions} object or list of objects}

\item{childLink}{Child link for an account entry}

\item{created}{Time the account was created}

\item{id}{Account ID}

\item{name}{Account name}

\item{permissions}{Permissions the user has for this account}

\item{selfLink}{Link for this account}

\item{starred}{Indicates whether this account is starred or not}

\item{updated}{Time the account was last modified}
}
\value{
Account object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:25:04
filename: /Users/mark/dev/R/autoGoogleAPI/googleanalyticsv3.auto/R/analytics_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Account Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for Analytics account entry.
}
\seealso{
Other Account functions: \code{\link{Account.childLink}},
  \code{\link{Account.permissions}}
}
