% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{AccountRef}
\alias{AccountRef}
\title{AccountRef Object}
\usage{
AccountRef(href = NULL, id = NULL, name = NULL)
}
\arguments{
\item{href}{Link for this account}

\item{id}{Account ID}

\item{name}{Account name}
}
\value{
AccountRef object
}
\description{
AccountRef Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for a linked account.
}
