% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Accounts}
\alias{Accounts}
\title{Accounts Object}
\usage{
Accounts(items = NULL, itemsPerPage = NULL, nextLink = NULL,
  previousLink = NULL, startIndex = NULL, totalResults = NULL,
  username = NULL)
}
\arguments{
\item{items}{A list of accounts}

\item{itemsPerPage}{The maximum number of entries the response can contain, regardless of the actual number of entries returned}

\item{nextLink}{Next link for this account collection}

\item{previousLink}{Previous link for this account collection}

\item{startIndex}{The starting index of the entries, which is 1 by default or otherwise specified by the start-index query parameter}

\item{totalResults}{The total number of results for the query, regardless of the number of results in the response}

\item{username}{Email ID of the authenticated user}
}
\value{
Accounts object
}
\description{
Accounts Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An account collection provides a list of Analytics accounts to which a user has access. The account collection is the entry point to all management information. Each resource in the collection corresponds to a single Analytics account.
}
