% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{AccountSummary}
\alias{AccountSummary}
\title{AccountSummary Object}
\usage{
AccountSummary(id = NULL, name = NULL, starred = NULL,
  webProperties = NULL)
}
\arguments{
\item{id}{Account ID}

\item{name}{Account name}

\item{starred}{Indicates whether this account is starred or not}

\item{webProperties}{List of web properties under this account}
}
\value{
AccountSummary object
}
\description{
AccountSummary Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for an Analytics AccountSummary. An AccountSummary is a lightweight tree comprised of properties/profiles.
}
