% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{AccountTicket}
\alias{AccountTicket}
\title{AccountTicket Object}
\usage{
AccountTicket(account = NULL, id = NULL, profile = NULL,
  redirectUri = NULL, webproperty = NULL)
}
\arguments{
\item{account}{Account for this ticket}

\item{id}{Account ticket ID used to access the account ticket}

\item{profile}{View (Profile) for the account}

\item{redirectUri}{Redirect URI where the user will be sent after accepting Terms of Service}

\item{webproperty}{Web property for the account}
}
\value{
AccountTicket object
}
\description{
AccountTicket Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for an Analytics account ticket. The account ticket consists of the ticket ID and the basic information for the account, property and profile.
}
\seealso{
Other AccountTicket functions: \code{\link{provisioning.createAccountTicket}}
}

