% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{AdWordsAccount}
\alias{AdWordsAccount}
\title{AdWordsAccount Object}
\usage{
AdWordsAccount(autoTaggingEnabled = NULL, customerId = NULL)
}
\arguments{
\item{autoTaggingEnabled}{True if auto-tagging is enabled on the AdWords account}

\item{customerId}{Customer ID}
}
\value{
AdWordsAccount object
}
\description{
AdWordsAccount Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for an AdWords account.
}
