% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\docType{package}
\name{analytics_googleAuthR}
\alias{analytics_googleAuthR}
\alias{analytics_googleAuthR-package}
\title{Google Analytics API
Views and manages your Google Analytics data.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-04 00:00:27
filename: /Users/mark/dev/R/autoGoogleAPI/googleanalyticsv3.auto/R/analytics_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/analytics
\item https://www.googleapis.com/auth/analytics.edit
\item https://www.googleapis.com/auth/analytics.manage.users
\item https://www.googleapis.com/auth/analytics.manage.users.readonly
\item https://www.googleapis.com/auth/analytics.provision
\item https://www.googleapis.com/auth/analytics.readonly
}
}

