% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{AnalyticsDataimportDeleteUploadDataRequest}
\alias{AnalyticsDataimportDeleteUploadDataRequest}
\title{AnalyticsDataimportDeleteUploadDataRequest Object}
\usage{
AnalyticsDataimportDeleteUploadDataRequest(customDataImportUids = NULL)
}
\arguments{
\item{customDataImportUids}{A list of upload UIDs}
}
\value{
AnalyticsDataimportDeleteUploadDataRequest object
}
\description{
AnalyticsDataimportDeleteUploadDataRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request template for the delete upload data request.
}
\seealso{
Other AnalyticsDataimportDeleteUploadDataRequest functions: \code{\link{management.uploads.deleteUploadData}}
}
