% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Column.attributes}
\alias{Column.attributes}
\title{Column.attributes Object}
\usage{
Column.attributes()
}
\value{
Column.attributes object
}
\description{
Column.attributes Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Map of attribute name and value for this column.
}
\seealso{
Other Column functions: \code{\link{Column}}
}

