% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Column}
\alias{Column}
\title{Column Object}
\usage{
Column(Column.attributes = NULL, attributes = NULL, id = NULL)
}
\arguments{
\item{Column.attributes}{The \link{Column.attributes} object or list of objects}

\item{attributes}{Map of attribute name and value for this column}

\item{id}{Column id}
}
\value{
Column object
}
\description{
Column Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for a metadata column.
}
\seealso{
Other Column functions: \code{\link{Column.attributes}}
}

