% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Columns}
\alias{Columns}
\title{Columns Object}
\usage{
Columns(attributeNames = NULL, etag = NULL, items = NULL,
  totalResults = NULL)
}
\arguments{
\item{attributeNames}{List of attributes names returned by columns}

\item{etag}{Etag of collection}

\item{items}{List of columns for a report type}

\item{totalResults}{Total number of columns returned in the response}
}
\value{
Columns object
}
\description{
Columns Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Lists columns (dimensions and metrics) for a particular report type.
}

