% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{CustomDataSource.childLink}
\alias{CustomDataSource.childLink}
\title{CustomDataSource.childLink Object}
\usage{
CustomDataSource.childLink(href = NULL, type = NULL)
}
\arguments{
\item{href}{Link to the list of daily uploads for this custom data source}

\item{type}{Value is 'analytics#dailyUploads'}
}
\value{
CustomDataSource.childLink object
}
\description{
CustomDataSource.childLink Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other CustomDataSource functions: \code{\link{CustomDataSource.parentLink}},
  \code{\link{CustomDataSource}}
}

