% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{CustomDataSource.parentLink}
\alias{CustomDataSource.parentLink}
\title{CustomDataSource.parentLink Object}
\usage{
CustomDataSource.parentLink(href = NULL, type = NULL)
}
\arguments{
\item{href}{Link to the web property to which this custom data source belongs}

\item{type}{Value is 'analytics#webproperty'}
}
\value{
CustomDataSource.parentLink object
}
\description{
CustomDataSource.parentLink Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Parent link for this custom data source. Points to the web property to which this custom data source belongs.
}
\seealso{
Other CustomDataSource functions: \code{\link{CustomDataSource.childLink}},
  \code{\link{CustomDataSource}}
}
