% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{CustomDataSource}
\alias{CustomDataSource}
\title{CustomDataSource Object}
\usage{
CustomDataSource(CustomDataSource.childLink = NULL,
  CustomDataSource.parentLink = NULL, accountId = NULL, childLink = NULL,
  created = NULL, description = NULL, id = NULL, importBehavior = NULL,
  name = NULL, parentLink = NULL, profilesLinked = NULL,
  selfLink = NULL, type = NULL, updated = NULL, uploadType = NULL,
  webPropertyId = NULL)
}
\arguments{
\item{CustomDataSource.childLink}{The \link{CustomDataSource.childLink} object or list of objects}

\item{CustomDataSource.parentLink}{The \link{CustomDataSource.parentLink} object or list of objects}

\item{accountId}{Account ID to which this custom data source belongs}

\item{childLink}{No description}

\item{created}{Time this custom data source was created}

\item{description}{Description of custom data source}

\item{id}{Custom data source ID}

\item{importBehavior}{No description}

\item{name}{Name of this custom data source}

\item{parentLink}{Parent link for this custom data source}

\item{profilesLinked}{IDs of views (profiles) linked to the custom data source}

\item{selfLink}{Link for this Analytics custom data source}

\item{type}{Type of the custom data source}

\item{updated}{Time this custom data source was last modified}

\item{uploadType}{No description}

\item{webPropertyId}{Web property ID of the form UA-XXXXX-YY to which this custom data source belongs}
}
\value{
CustomDataSource object
}
\description{
CustomDataSource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for an Analytics custom data source.
}
\seealso{
Other CustomDataSource functions: \code{\link{CustomDataSource.childLink}},
  \code{\link{CustomDataSource.parentLink}}
}
