% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{CustomDataSources}
\alias{CustomDataSources}
\title{CustomDataSources Object}
\usage{
CustomDataSources(items = NULL, itemsPerPage = NULL, nextLink = NULL,
  previousLink = NULL, startIndex = NULL, totalResults = NULL,
  username = NULL)
}
\arguments{
\item{items}{Collection of custom data sources}

\item{itemsPerPage}{The maximum number of resources the response can contain, regardless of the actual number of resources returned}

\item{nextLink}{Link to next page for this custom data source collection}

\item{previousLink}{Link to previous page for this custom data source collection}

\item{startIndex}{The starting index of the resources, which is 1 by default or otherwise specified by the start-index query parameter}

\item{totalResults}{The total number of results for the query, regardless of the number of results in the response}

\item{username}{Email ID of the authenticated user}
}
\value{
CustomDataSources object
}
\description{
CustomDataSources Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Lists Analytics custom data sources to which the user has access. Each resource in the collection corresponds to a single Analytics custom data source.
}
