% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{CustomDimension.parentLink}
\alias{CustomDimension.parentLink}
\title{CustomDimension.parentLink Object}
\usage{
CustomDimension.parentLink(href = NULL, type = NULL)
}
\arguments{
\item{href}{Link to the property to which the custom dimension belongs}

\item{type}{Type of the parent link}
}
\value{
CustomDimension.parentLink object
}
\description{
CustomDimension.parentLink Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Parent link for the custom dimension. Points to the property to which the custom dimension belongs.
}
\seealso{
Other CustomDimension functions: \code{\link{CustomDimension}},
  \code{\link{management.customDimensions.insert}},
  \code{\link{management.customDimensions.patch}},
  \code{\link{management.customDimensions.update}}
}

