% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{CustomDimension}
\alias{CustomDimension}
\title{CustomDimension Object}
\usage{
CustomDimension(CustomDimension.parentLink = NULL, accountId = NULL,
  active = NULL, id = NULL, name = NULL, parentLink = NULL,
  scope = NULL, webPropertyId = NULL)
}
\arguments{
\item{CustomDimension.parentLink}{The \link{CustomDimension.parentLink} object or list of objects}

\item{accountId}{Account ID}

\item{active}{Boolean indicating whether the custom dimension is active}

\item{id}{Custom dimension ID}

\item{name}{Name of the custom dimension}

\item{parentLink}{Parent link for the custom dimension}

\item{scope}{Scope of the custom dimension: HIT, SESSION, USER or PRODUCT}

\item{webPropertyId}{Property ID}
}
\value{
CustomDimension object
}
\description{
CustomDimension Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for Analytics Custom Dimension.
}
\seealso{
Other CustomDimension functions: \code{\link{CustomDimension.parentLink}},
  \code{\link{management.customDimensions.insert}},
  \code{\link{management.customDimensions.patch}},
  \code{\link{management.customDimensions.update}}
}
