% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{CustomMetric.parentLink}
\alias{CustomMetric.parentLink}
\title{CustomMetric.parentLink Object}
\usage{
CustomMetric.parentLink(href = NULL, type = NULL)
}
\arguments{
\item{href}{Link to the property to which the custom metric belongs}

\item{type}{Type of the parent link}
}
\value{
CustomMetric.parentLink object
}
\description{
CustomMetric.parentLink Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Parent link for the custom metric. Points to the property to which the custom metric belongs.
}
\seealso{
Other CustomMetric functions: \code{\link{CustomMetric}},
  \code{\link{management.customMetrics.insert}},
  \code{\link{management.customMetrics.patch}},
  \code{\link{management.customMetrics.update}}
}

