% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{CustomMetric}
\alias{CustomMetric}
\title{CustomMetric Object}
\usage{
CustomMetric(CustomMetric.parentLink = NULL, accountId = NULL,
  active = NULL, id = NULL, max_value = NULL, min_value = NULL,
  name = NULL, parentLink = NULL, scope = NULL, type = NULL,
  webPropertyId = NULL)
}
\arguments{
\item{CustomMetric.parentLink}{The \link{CustomMetric.parentLink} object or list of objects}

\item{accountId}{Account ID}

\item{active}{Boolean indicating whether the custom metric is active}

\item{id}{Custom metric ID}

\item{max_value}{Max value of custom metric}

\item{min_value}{Min value of custom metric}

\item{name}{Name of the custom metric}

\item{parentLink}{Parent link for the custom metric}

\item{scope}{Scope of the custom metric: HIT or PRODUCT}

\item{type}{Data type of custom metric}

\item{webPropertyId}{Property ID}
}
\value{
CustomMetric object
}
\description{
CustomMetric Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for Analytics Custom Metric.
}
\seealso{
Other CustomMetric functions: \code{\link{CustomMetric.parentLink}},
  \code{\link{management.customMetrics.insert}},
  \code{\link{management.customMetrics.patch}},
  \code{\link{management.customMetrics.update}}
}
