% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{data.realtime.get}
\alias{data.realtime.get}
\title{Returns real time data for a view (profile).}
\usage{
data.realtime.get(ids, metrics, dimensions = NULL, filters = NULL,
  max.results = NULL, sort = NULL)
}
\arguments{
\item{ids}{Unique table ID for retrieving real time data}

\item{metrics}{A comma-separated list of real time metrics}

\item{dimensions}{A comma-separated list of real time dimensions}

\item{filters}{A comma-separated list of dimension or metric filters to be applied to real time data}

\item{max.results}{The maximum number of entries to include in this feed}

\item{sort}{A comma-separated list of dimensions or metrics that determine the sort order for real time data}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics
\item https://www.googleapis.com/auth/analytics.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics, https://www.googleapis.com/auth/analytics.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}
}

