% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{EntityAdWordsLink.entity}
\alias{EntityAdWordsLink.entity}
\title{EntityAdWordsLink.entity Object}
\usage{
EntityAdWordsLink.entity(webPropertyRef = NULL)
}
\arguments{
\item{webPropertyRef}{No description}
}
\value{
EntityAdWordsLink.entity object
}
\description{
EntityAdWordsLink.entity Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Web property being linked.
}
\seealso{
Other EntityAdWordsLink functions: \code{\link{EntityAdWordsLink}},
  \code{\link{management.webPropertyAdWordsLinks.insert}},
  \code{\link{management.webPropertyAdWordsLinks.patch}},
  \code{\link{management.webPropertyAdWordsLinks.update}}
}
