% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{EntityAdWordsLink}
\alias{EntityAdWordsLink}
\title{EntityAdWordsLink Object}
\usage{
EntityAdWordsLink(EntityAdWordsLink.entity = NULL, adWordsAccounts = NULL,
  entity = NULL, id = NULL, name = NULL, profileIds = NULL,
  selfLink = NULL)
}
\arguments{
\item{EntityAdWordsLink.entity}{The \link{EntityAdWordsLink.entity} object or list of objects}

\item{adWordsAccounts}{A list of AdWords client accounts}

\item{entity}{Web property being linked}

\item{id}{Entity AdWords link ID}

\item{name}{Name of the link}

\item{profileIds}{IDs of linked Views (Profiles) represented as strings}

\item{selfLink}{URL link for this Google Analytics - Google AdWords link}
}
\value{
EntityAdWordsLink object
}
\description{
EntityAdWordsLink Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for Analytics Entity AdWords Link.
}
\seealso{
Other EntityAdWordsLink functions: \code{\link{EntityAdWordsLink.entity}},
  \code{\link{management.webPropertyAdWordsLinks.insert}},
  \code{\link{management.webPropertyAdWordsLinks.patch}},
  \code{\link{management.webPropertyAdWordsLinks.update}}
}

