% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{EntityAdWordsLinks}
\alias{EntityAdWordsLinks}
\title{EntityAdWordsLinks Object}
\usage{
EntityAdWordsLinks(items = NULL, itemsPerPage = NULL, nextLink = NULL,
  previousLink = NULL, startIndex = NULL, totalResults = NULL)
}
\arguments{
\item{items}{A list of entity AdWords links}

\item{itemsPerPage}{The maximum number of entries the response can contain, regardless of the actual number of entries returned}

\item{nextLink}{Next link for this AdWords link collection}

\item{previousLink}{Previous link for this AdWords link collection}

\item{startIndex}{The starting index of the entries, which is 1 by default or otherwise specified by the start-index query parameter}

\item{totalResults}{The total number of results for the query, regardless of the number of results in the response}
}
\value{
EntityAdWordsLinks object
}
\description{
EntityAdWordsLinks Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An entity AdWords link collection provides a list of GA-AdWords links Each resource in this collection corresponds to a single link.
}

