% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{EntityUserLink.entity}
\alias{EntityUserLink.entity}
\title{EntityUserLink.entity Object}
\usage{
EntityUserLink.entity(accountRef = NULL, profileRef = NULL,
  webPropertyRef = NULL)
}
\arguments{
\item{accountRef}{Account for this link}

\item{profileRef}{View (Profile) for this link}

\item{webPropertyRef}{Web property for this link}
}
\value{
EntityUserLink.entity object
}
\description{
EntityUserLink.entity Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Entity for this link. It can be an account, a web property, or a view (profile).
}
\seealso{
Other EntityUserLink functions: \code{\link{EntityUserLink.permissions}},
  \code{\link{EntityUserLink}},
  \code{\link{management.accountUserLinks.insert}},
  \code{\link{management.accountUserLinks.update}},
  \code{\link{management.profileUserLinks.insert}},
  \code{\link{management.profileUserLinks.update}},
  \code{\link{management.webpropertyUserLinks.insert}},
  \code{\link{management.webpropertyUserLinks.update}}
}

