% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{EntityUserLink.permissions}
\alias{EntityUserLink.permissions}
\title{EntityUserLink.permissions Object}
\usage{
EntityUserLink.permissions(local = NULL)
}
\arguments{
\item{local}{Permissions that a user has been assigned at this very level}
}
\value{
EntityUserLink.permissions object
}
\description{
EntityUserLink.permissions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Permissions the user has for this entity.
}
\seealso{
Other EntityUserLink functions: \code{\link{EntityUserLink.entity}},
  \code{\link{EntityUserLink}},
  \code{\link{management.accountUserLinks.insert}},
  \code{\link{management.accountUserLinks.update}},
  \code{\link{management.profileUserLinks.insert}},
  \code{\link{management.profileUserLinks.update}},
  \code{\link{management.webpropertyUserLinks.insert}},
  \code{\link{management.webpropertyUserLinks.update}}
}
