% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{EntityUserLink}
\alias{EntityUserLink}
\title{EntityUserLink Object}
\usage{
EntityUserLink(EntityUserLink.entity = NULL,
  EntityUserLink.permissions = NULL, entity = NULL, id = NULL,
  permissions = NULL, selfLink = NULL, userRef = NULL)
}
\arguments{
\item{EntityUserLink.entity}{The \link{EntityUserLink.entity} object or list of objects}

\item{EntityUserLink.permissions}{The \link{EntityUserLink.permissions} object or list of objects}

\item{entity}{Entity for this link}

\item{id}{Entity user link ID}

\item{permissions}{Permissions the user has for this entity}

\item{selfLink}{Self link for this resource}

\item{userRef}{User reference}
}
\value{
EntityUserLink object
}
\description{
EntityUserLink Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for an Analytics Entity-User Link. Returns permissions that a user has for an entity.
}
\seealso{
Other EntityUserLink functions: \code{\link{EntityUserLink.entity}},
  \code{\link{EntityUserLink.permissions}},
  \code{\link{management.accountUserLinks.insert}},
  \code{\link{management.accountUserLinks.update}},
  \code{\link{management.profileUserLinks.insert}},
  \code{\link{management.profileUserLinks.update}},
  \code{\link{management.webpropertyUserLinks.insert}},
  \code{\link{management.webpropertyUserLinks.update}}
}

