% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{EntityUserLinks}
\alias{EntityUserLinks}
\title{EntityUserLinks Object}
\usage{
EntityUserLinks(items = NULL, itemsPerPage = NULL, nextLink = NULL,
  previousLink = NULL, startIndex = NULL, totalResults = NULL)
}
\arguments{
\item{items}{A list of entity user links}

\item{itemsPerPage}{The maximum number of entries the response can contain, regardless of the actual number of entries returned}

\item{nextLink}{Next link for this account collection}

\item{previousLink}{Previous link for this account collection}

\item{startIndex}{The starting index of the entries, which is 1 by default or otherwise specified by the start-index query parameter}

\item{totalResults}{The total number of results for the query, regardless of the number of results in the response}
}
\value{
EntityUserLinks object
}
\description{
EntityUserLinks Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An entity user link collection provides a list of Analytics ACL links Each resource in this collection corresponds to a single link.
}
