% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Experiment.parentLink}
\alias{Experiment.parentLink}
\title{Experiment.parentLink Object}
\usage{
Experiment.parentLink(href = NULL, type = NULL)
}
\arguments{
\item{href}{Link to the view (profile) to which this experiment belongs}

\item{type}{Value is 'analytics#profile'}
}
\value{
Experiment.parentLink object
}
\description{
Experiment.parentLink Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Parent link for an experiment. Points to the view (profile) to which this experiment belongs.
}
\seealso{
Other Experiment functions: \code{\link{Experiment.variations}},
  \code{\link{Experiment}},
  \code{\link{management.experiments.insert}},
  \code{\link{management.experiments.patch}},
  \code{\link{management.experiments.update}}
}
