% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Experiment}
\alias{Experiment}
\title{Experiment Object}
\usage{
Experiment(Experiment.parentLink = NULL, Experiment.variations = NULL,
  accountId = NULL, created = NULL, description = NULL,
  editableInGaUi = NULL, endTime = NULL, equalWeighting = NULL,
  id = NULL, internalWebPropertyId = NULL,
  minimumExperimentLengthInDays = NULL, name = NULL,
  objectiveMetric = NULL, optimizationType = NULL, parentLink = NULL,
  profileId = NULL, reasonExperimentEnded = NULL,
  rewriteVariationUrlsAsOriginal = NULL, selfLink = NULL,
  servingFramework = NULL, snippet = NULL, startTime = NULL,
  status = NULL, trafficCoverage = NULL, updated = NULL,
  variations = NULL, webPropertyId = NULL, winnerConfidenceLevel = NULL,
  winnerFound = NULL)
}
\arguments{
\item{Experiment.parentLink}{The \link{Experiment.parentLink} object or list of objects}

\item{Experiment.variations}{The \link{Experiment.variations} object or list of objects}

\item{accountId}{Account ID to which this experiment belongs}

\item{created}{Time the experiment was created}

\item{description}{Notes about this experiment}

\item{editableInGaUi}{If true, the end user will be able to edit the experiment via the Google Analytics user interface}

\item{endTime}{The ending time of the experiment (the time the status changed from RUNNING to ENDED)}

\item{equalWeighting}{Boolean specifying whether to distribute traffic evenly across all variations}

\item{id}{Experiment ID}

\item{internalWebPropertyId}{Internal ID for the web property to which this experiment belongs}

\item{minimumExperimentLengthInDays}{An integer number in [3, 90]}

\item{name}{Experiment name}

\item{objectiveMetric}{The metric that the experiment is optimizing}

\item{optimizationType}{Whether the objectiveMetric should be minimized or maximized}

\item{parentLink}{Parent link for an experiment}

\item{profileId}{View (Profile) ID to which this experiment belongs}

\item{reasonExperimentEnded}{Why the experiment ended}

\item{rewriteVariationUrlsAsOriginal}{Boolean specifying whether variations URLS are rewritten to match those of the original}

\item{selfLink}{Link for this experiment}

\item{servingFramework}{The framework used to serve the experiment variations and evaluate the results}

\item{snippet}{The snippet of code to include on the control page(s)}

\item{startTime}{The starting time of the experiment (the time the status changed from READY_TO_RUN to RUNNING)}

\item{status}{Experiment status}

\item{trafficCoverage}{A floating-point number in (0, 1]}

\item{updated}{Time the experiment was last modified}

\item{variations}{Array of variations}

\item{webPropertyId}{Web property ID to which this experiment belongs}

\item{winnerConfidenceLevel}{A floating-point number in (0, 1)}

\item{winnerFound}{Boolean specifying whether a winner has been found for this experiment}
}
\value{
Experiment object
}
\description{
Experiment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for Analytics experiment resource.
}
\seealso{
Other Experiment functions: \code{\link{Experiment.parentLink}},
  \code{\link{Experiment.variations}},
  \code{\link{management.experiments.insert}},
  \code{\link{management.experiments.patch}},
  \code{\link{management.experiments.update}}
}
