% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Experiment.variations}
\alias{Experiment.variations}
\title{Experiment.variations Object}
\usage{
Experiment.variations()
}
\value{
Experiment.variations object
}
\description{
Experiment.variations Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Experiment functions: \code{\link{Experiment.parentLink}},
  \code{\link{Experiment}},
  \code{\link{management.experiments.insert}},
  \code{\link{management.experiments.patch}},
  \code{\link{management.experiments.update}}
}
