% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Experiments}
\alias{Experiments}
\title{Experiments Object}
\usage{
Experiments(items = NULL, itemsPerPage = NULL, nextLink = NULL,
  previousLink = NULL, startIndex = NULL, totalResults = NULL,
  username = NULL)
}
\arguments{
\item{items}{A list of experiments}

\item{itemsPerPage}{The maximum number of resources the response can contain, regardless of the actual number of resources returned}

\item{nextLink}{Link to next page for this experiment collection}

\item{previousLink}{Link to previous page for this experiment collection}

\item{startIndex}{The starting index of the resources, which is 1 by default or otherwise specified by the start-index query parameter}

\item{totalResults}{The total number of results for the query, regardless of the number of resources in the result}

\item{username}{Email ID of the authenticated user}
}
\value{
Experiments object
}
\description{
Experiments Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An experiment collection lists Analytics experiments to which the user has access. Each view (profile) can have a set of experiments. Each resource in the Experiment collection corresponds to a single Analytics experiment.
}
