% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Filter.parentLink}
\alias{Filter.parentLink}
\title{Filter.parentLink Object}
\usage{
Filter.parentLink(href = NULL, type = NULL)
}
\arguments{
\item{href}{Link to the account to which this filter belongs}

\item{type}{Value is 'analytics#account'}
}
\value{
Filter.parentLink object
}
\description{
Filter.parentLink Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Parent link for this filter. Points to the account to which this filter belongs.
}
\seealso{
Other Filter functions: \code{\link{Filter.advancedDetails}},
  \code{\link{Filter.lowercaseDetails}},
  \code{\link{Filter.searchAndReplaceDetails}},
  \code{\link{Filter.uppercaseDetails}},
  \code{\link{Filter}},
  \code{\link{management.filters.insert}},
  \code{\link{management.filters.patch}},
  \code{\link{management.filters.update}}
}
