% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Filter}
\alias{Filter}
\title{Filter Object}
\usage{
Filter(Filter.advancedDetails = NULL, Filter.lowercaseDetails = NULL,
  Filter.parentLink = NULL, Filter.searchAndReplaceDetails = NULL,
  Filter.uppercaseDetails = NULL, accountId = NULL,
  advancedDetails = NULL, excludeDetails = NULL, id = NULL,
  includeDetails = NULL, lowercaseDetails = NULL, name = NULL,
  parentLink = NULL, searchAndReplaceDetails = NULL, type = NULL,
  uppercaseDetails = NULL)
}
\arguments{
\item{Filter.advancedDetails}{The \link{Filter.advancedDetails} object or list of objects}

\item{Filter.lowercaseDetails}{The \link{Filter.lowercaseDetails} object or list of objects}

\item{Filter.parentLink}{The \link{Filter.parentLink} object or list of objects}

\item{Filter.searchAndReplaceDetails}{The \link{Filter.searchAndReplaceDetails} object or list of objects}

\item{Filter.uppercaseDetails}{The \link{Filter.uppercaseDetails} object or list of objects}

\item{accountId}{Account ID to which this filter belongs}

\item{advancedDetails}{Details for the filter of the type ADVANCED}

\item{excludeDetails}{Details for the filter of the type EXCLUDE}

\item{id}{Filter ID}

\item{includeDetails}{Details for the filter of the type INCLUDE}

\item{lowercaseDetails}{Details for the filter of the type LOWER}

\item{name}{Name of this filter}

\item{parentLink}{Parent link for this filter}

\item{searchAndReplaceDetails}{Details for the filter of the type SEARCH_AND_REPLACE}

\item{type}{Type of this filter}

\item{uppercaseDetails}{Details for the filter of the type UPPER}
}
\value{
Filter object
}
\description{
Filter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for an Analytics account filter.
}
\seealso{
Other Filter functions: \code{\link{Filter.advancedDetails}},
  \code{\link{Filter.lowercaseDetails}},
  \code{\link{Filter.parentLink}},
  \code{\link{Filter.searchAndReplaceDetails}},
  \code{\link{Filter.uppercaseDetails}},
  \code{\link{management.filters.insert}},
  \code{\link{management.filters.patch}},
  \code{\link{management.filters.update}}
}
