% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Filter.searchAndReplaceDetails}
\alias{Filter.searchAndReplaceDetails}
\title{Filter.searchAndReplaceDetails Object}
\usage{
Filter.searchAndReplaceDetails(caseSensitive = NULL, field = NULL,
  fieldIndex = NULL, replaceString = NULL, searchString = NULL)
}
\arguments{
\item{caseSensitive}{Determines if the filter is case sensitive}

\item{field}{Field to use in the filter}

\item{fieldIndex}{The Index of the custom dimension}

\item{replaceString}{Term to replace the search term with}

\item{searchString}{Term to search}
}
\value{
Filter.searchAndReplaceDetails object
}
\description{
Filter.searchAndReplaceDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details for the filter of the type SEARCH_AND_REPLACE.
}
\seealso{
Other Filter functions: \code{\link{Filter.advancedDetails}},
  \code{\link{Filter.lowercaseDetails}},
  \code{\link{Filter.parentLink}},
  \code{\link{Filter.uppercaseDetails}},
  \code{\link{Filter}},
  \code{\link{management.filters.insert}},
  \code{\link{management.filters.patch}},
  \code{\link{management.filters.update}}
}
