% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Filter.uppercaseDetails}
\alias{Filter.uppercaseDetails}
\title{Filter.uppercaseDetails Object}
\usage{
Filter.uppercaseDetails(field = NULL, fieldIndex = NULL)
}
\arguments{
\item{field}{Field to use in the filter}

\item{fieldIndex}{The Index of the custom dimension}
}
\value{
Filter.uppercaseDetails object
}
\description{
Filter.uppercaseDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details for the filter of the type UPPER.
}
\seealso{
Other Filter functions: \code{\link{Filter.advancedDetails}},
  \code{\link{Filter.lowercaseDetails}},
  \code{\link{Filter.parentLink}},
  \code{\link{Filter.searchAndReplaceDetails}},
  \code{\link{Filter}},
  \code{\link{management.filters.insert}},
  \code{\link{management.filters.patch}},
  \code{\link{management.filters.update}}
}
