% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{FilterExpression}
\alias{FilterExpression}
\title{FilterExpression Object}
\usage{
FilterExpression(caseSensitive = NULL, expressionValue = NULL,
  field = NULL, fieldIndex = NULL, matchType = NULL)
}
\arguments{
\item{caseSensitive}{Determines if the filter is case sensitive}

\item{expressionValue}{Filter expression value}

\item{field}{Field to filter}

\item{fieldIndex}{The Index of the custom dimension}

\item{matchType}{Match type for this filter}
}
\value{
FilterExpression object
}
\description{
FilterExpression Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for an Analytics filter expression.
}

