% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{FilterRef}
\alias{FilterRef}
\title{FilterRef Object}
\usage{
FilterRef(href = NULL, id = NULL)
}
\arguments{
\item{href}{Link for this filter}

\item{id}{Filter ID}
}
\value{
FilterRef object
}
\description{
FilterRef Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for a profile filter link.
}

