% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Filters}
\alias{Filters}
\title{Filters Object}
\usage{
Filters(items = NULL, itemsPerPage = NULL, nextLink = NULL,
  previousLink = NULL, startIndex = NULL, totalResults = NULL,
  username = NULL)
}
\arguments{
\item{items}{A list of filters}

\item{itemsPerPage}{The maximum number of resources the response can contain, regardless of the actual number of resources returned}

\item{nextLink}{Link to next page for this filter collection}

\item{previousLink}{Link to previous page for this filter collection}

\item{startIndex}{The starting index of the resources, which is 1 by default or otherwise specified by the start-index query parameter}

\item{totalResults}{The total number of results for the query, regardless of the number of results in the response}

\item{username}{Email ID of the authenticated user}
}
\value{
Filters object
}
\description{
Filters Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A filter collection lists filters created by users in an Analytics account. Each resource in the collection corresponds to a filter.
}
