% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{GaData.columnHeaders}
\alias{GaData.columnHeaders}
\title{GaData.columnHeaders Object}
\usage{
GaData.columnHeaders()
}
\value{
GaData.columnHeaders object
}
\description{
GaData.columnHeaders Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other GaData functions: \code{\link{GaData.dataTable.cols}},
  \code{\link{GaData.dataTable.rows.c}},
  \code{\link{GaData.dataTable.rows}},
  \code{\link{GaData.dataTable}},
  \code{\link{GaData.profileInfo}},
  \code{\link{GaData.query}},
  \code{\link{GaData.totalsForAllResults}},
  \code{\link{GaData}}
}

