% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{GaData.dataTable.cols}
\alias{GaData.dataTable.cols}
\title{GaData.dataTable.cols Object}
\usage{
GaData.dataTable.cols()
}
\value{
GaData.dataTable.cols object
}
\description{
GaData.dataTable.cols Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other GaData functions: \code{\link{GaData.columnHeaders}},
  \code{\link{GaData.dataTable.rows.c}},
  \code{\link{GaData.dataTable.rows}},
  \code{\link{GaData.dataTable}},
  \code{\link{GaData.profileInfo}},
  \code{\link{GaData.query}},
  \code{\link{GaData.totalsForAllResults}},
  \code{\link{GaData}}
}

