% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{GaData.dataTable}
\alias{GaData.dataTable}
\title{GaData.dataTable Object}
\usage{
GaData.dataTable(GaData.dataTable.cols = NULL, GaData.dataTable.rows = NULL,
  GaData.dataTable.rows.c = NULL, cols = NULL, rows = NULL)
}
\arguments{
\item{GaData.dataTable.cols}{The \link{GaData.dataTable.cols} object or list of objects}

\item{GaData.dataTable.rows}{The \link{GaData.dataTable.rows} object or list of objects}

\item{GaData.dataTable.rows.c}{The \link{GaData.dataTable.rows.c} object or list of objects}

\item{cols}{No description}

\item{rows}{No description}
}
\value{
GaData.dataTable object
}
\description{
GaData.dataTable Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other GaData functions: \code{\link{GaData.columnHeaders}},
  \code{\link{GaData.dataTable.cols}},
  \code{\link{GaData.dataTable.rows.c}},
  \code{\link{GaData.dataTable.rows}},
  \code{\link{GaData.profileInfo}},
  \code{\link{GaData.query}},
  \code{\link{GaData.totalsForAllResults}},
  \code{\link{GaData}}
}
