% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{GaData.dataTable.rows.c}
\alias{GaData.dataTable.rows.c}
\title{GaData.dataTable.rows.c Object}
\usage{
GaData.dataTable.rows.c()
}
\value{
GaData.dataTable.rows.c object
}
\description{
GaData.dataTable.rows.c Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other GaData functions: \code{\link{GaData.columnHeaders}},
  \code{\link{GaData.dataTable.cols}},
  \code{\link{GaData.dataTable.rows}},
  \code{\link{GaData.dataTable}},
  \code{\link{GaData.profileInfo}},
  \code{\link{GaData.query}},
  \code{\link{GaData.totalsForAllResults}},
  \code{\link{GaData}}
}

