% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{GaData.dataTable.rows}
\alias{GaData.dataTable.rows}
\title{GaData.dataTable.rows Object}
\usage{
GaData.dataTable.rows(GaData.dataTable.rows.c = NULL)
}
\arguments{
\item{GaData.dataTable.rows.c}{The \link{GaData.dataTable.rows.c} object or list of objects}
}
\value{
GaData.dataTable.rows object
}
\description{
GaData.dataTable.rows Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other GaData functions: \code{\link{GaData.columnHeaders}},
  \code{\link{GaData.dataTable.cols}},
  \code{\link{GaData.dataTable.rows.c}},
  \code{\link{GaData.dataTable}},
  \code{\link{GaData.profileInfo}},
  \code{\link{GaData.query}},
  \code{\link{GaData.totalsForAllResults}},
  \code{\link{GaData}}
}
