% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{GaData.query}
\alias{GaData.query}
\title{GaData.query Object}
\usage{
GaData.query(dimensions = NULL, end.date = NULL, filters = NULL,
  ids = NULL, max.results = NULL, metrics = NULL, samplingLevel = NULL,
  segment = NULL, sort = NULL, start.date = NULL, start.index = NULL)
}
\arguments{
\item{dimensions}{List of analytics dimensions}

\item{end.date}{End date}

\item{filters}{Comma-separated list of dimension or metric filters}

\item{ids}{Unique table ID}

\item{max.results}{Maximum results per page}

\item{metrics}{List of analytics metrics}

\item{samplingLevel}{Desired sampling level}

\item{segment}{Analytics advanced segment}

\item{sort}{List of dimensions or metrics based on which Analytics data is sorted}

\item{start.date}{Start date}

\item{start.index}{Start index}
}
\value{
GaData.query object
}
\description{
GaData.query Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Analytics data request query parameters.
}
\seealso{
Other GaData functions: \code{\link{GaData.columnHeaders}},
  \code{\link{GaData.dataTable.cols}},
  \code{\link{GaData.dataTable.rows.c}},
  \code{\link{GaData.dataTable.rows}},
  \code{\link{GaData.dataTable}},
  \code{\link{GaData.profileInfo}},
  \code{\link{GaData.totalsForAllResults}},
  \code{\link{GaData}}
}
