% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{GaData}
\alias{GaData}
\title{GaData Object}
\usage{
GaData(GaData.columnHeaders = NULL, GaData.dataTable = NULL,
  GaData.dataTable.cols = NULL, GaData.dataTable.rows = NULL,
  GaData.dataTable.rows.c = NULL, GaData.profileInfo = NULL,
  GaData.query = NULL, GaData.totalsForAllResults = NULL,
  columnHeaders = NULL, containsSampledData = NULL,
  dataLastRefreshed = NULL, dataTable = NULL, id = NULL,
  itemsPerPage = NULL, nextLink = NULL, previousLink = NULL,
  profileInfo = NULL, query = NULL, rows = NULL, sampleSize = NULL,
  sampleSpace = NULL, selfLink = NULL, totalResults = NULL,
  totalsForAllResults = NULL)
}
\arguments{
\item{GaData.columnHeaders}{The \link{GaData.columnHeaders} object or list of objects}

\item{GaData.dataTable}{The \link{GaData.dataTable} object or list of objects}

\item{GaData.dataTable.cols}{The \link{GaData.dataTable.cols} object or list of objects}

\item{GaData.dataTable.rows}{The \link{GaData.dataTable.rows} object or list of objects}

\item{GaData.dataTable.rows.c}{The \link{GaData.dataTable.rows.c} object or list of objects}

\item{GaData.profileInfo}{The \link{GaData.profileInfo} object or list of objects}

\item{GaData.query}{The \link{GaData.query} object or list of objects}

\item{GaData.totalsForAllResults}{The \link{GaData.totalsForAllResults} object or list of objects}

\item{columnHeaders}{Column headers that list dimension names followed by the metric names}

\item{containsSampledData}{Determines if Analytics data contains samples}

\item{dataLastRefreshed}{The last refreshed time in seconds for Analytics data}

\item{dataTable}{No description}

\item{id}{Unique ID for this data response}

\item{itemsPerPage}{The maximum number of rows the response can contain, regardless of the actual number of rows returned}

\item{nextLink}{Link to next page for this Analytics data query}

\item{previousLink}{Link to previous page for this Analytics data query}

\item{profileInfo}{Information for the view (profile), for which the Analytics data was requested}

\item{query}{Analytics data request query parameters}

\item{rows}{Analytics data rows, where each row contains a list of dimension values followed by the metric values}

\item{sampleSize}{The number of samples used to calculate the result}

\item{sampleSpace}{Total size of the sample space from which the samples were selected}

\item{selfLink}{Link to this page}

\item{totalResults}{The total number of rows for the query, regardless of the number of rows in the response}

\item{totalsForAllResults}{Total values for the requested metrics over all the results, not just the results returned in this response}
}
\value{
GaData object
}
\description{
GaData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Analytics data for a given view (profile).
}
\seealso{
Other GaData functions: \code{\link{GaData.columnHeaders}},
  \code{\link{GaData.dataTable.cols}},
  \code{\link{GaData.dataTable.rows.c}},
  \code{\link{GaData.dataTable.rows}},
  \code{\link{GaData.dataTable}},
  \code{\link{GaData.profileInfo}},
  \code{\link{GaData.query}},
  \code{\link{GaData.totalsForAllResults}}
}
