% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{GaData.totalsForAllResults}
\alias{GaData.totalsForAllResults}
\title{GaData.totalsForAllResults Object}
\usage{
GaData.totalsForAllResults()
}
\value{
GaData.totalsForAllResults object
}
\description{
GaData.totalsForAllResults Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Total values for the requested metrics over all the results, not just the results returned in this response. The order of the metric totals is same as the metric order specified in the request.
}
\seealso{
Other GaData functions: \code{\link{GaData.columnHeaders}},
  \code{\link{GaData.dataTable.cols}},
  \code{\link{GaData.dataTable.rows.c}},
  \code{\link{GaData.dataTable.rows}},
  \code{\link{GaData.dataTable}},
  \code{\link{GaData.profileInfo}},
  \code{\link{GaData.query}}, \code{\link{GaData}}
}
