% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Goal.eventDetails.eventConditions}
\alias{Goal.eventDetails.eventConditions}
\title{Goal.eventDetails.eventConditions Object}
\usage{
Goal.eventDetails.eventConditions()
}
\value{
Goal.eventDetails.eventConditions object
}
\description{
Goal.eventDetails.eventConditions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Goal functions: \code{\link{Goal.eventDetails}},
  \code{\link{Goal.parentLink}},
  \code{\link{Goal.urlDestinationDetails.steps}},
  \code{\link{Goal.urlDestinationDetails}},
  \code{\link{Goal.visitNumPagesDetails}},
  \code{\link{Goal.visitTimeOnSiteDetails}},
  \code{\link{Goal}},
  \code{\link{management.goals.insert}},
  \code{\link{management.goals.patch}},
  \code{\link{management.goals.update}}
}

