% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Goal.eventDetails}
\alias{Goal.eventDetails}
\title{Goal.eventDetails Object}
\usage{
Goal.eventDetails(Goal.eventDetails.eventConditions = NULL,
  eventConditions = NULL, useEventValue = NULL)
}
\arguments{
\item{Goal.eventDetails.eventConditions}{The \link{Goal.eventDetails.eventConditions} object or list of objects}

\item{eventConditions}{List of event conditions}

\item{useEventValue}{Determines if the event value should be used as the value for this goal}
}
\value{
Goal.eventDetails object
}
\description{
Goal.eventDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details for the goal of the type EVENT.
}
\seealso{
Other Goal functions: \code{\link{Goal.eventDetails.eventConditions}},
  \code{\link{Goal.parentLink}},
  \code{\link{Goal.urlDestinationDetails.steps}},
  \code{\link{Goal.urlDestinationDetails}},
  \code{\link{Goal.visitNumPagesDetails}},
  \code{\link{Goal.visitTimeOnSiteDetails}},
  \code{\link{Goal}},
  \code{\link{management.goals.insert}},
  \code{\link{management.goals.patch}},
  \code{\link{management.goals.update}}
}
