% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Goal.parentLink}
\alias{Goal.parentLink}
\title{Goal.parentLink Object}
\usage{
Goal.parentLink(href = NULL, type = NULL)
}
\arguments{
\item{href}{Link to the view (profile) to which this goal belongs}

\item{type}{Value is 'analytics#profile'}
}
\value{
Goal.parentLink object
}
\description{
Goal.parentLink Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Parent link for a goal. Points to the view (profile) to which this goal belongs.
}
\seealso{
Other Goal functions: \code{\link{Goal.eventDetails.eventConditions}},
  \code{\link{Goal.eventDetails}},
  \code{\link{Goal.urlDestinationDetails.steps}},
  \code{\link{Goal.urlDestinationDetails}},
  \code{\link{Goal.visitNumPagesDetails}},
  \code{\link{Goal.visitTimeOnSiteDetails}},
  \code{\link{Goal}},
  \code{\link{management.goals.insert}},
  \code{\link{management.goals.patch}},
  \code{\link{management.goals.update}}
}
