% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Goal}
\alias{Goal}
\title{Goal Object}
\usage{
Goal(Goal.eventDetails = NULL, Goal.eventDetails.eventConditions = NULL,
  Goal.parentLink = NULL, Goal.urlDestinationDetails = NULL,
  Goal.urlDestinationDetails.steps = NULL, Goal.visitNumPagesDetails = NULL,
  Goal.visitTimeOnSiteDetails = NULL, accountId = NULL, active = NULL,
  created = NULL, eventDetails = NULL, id = NULL,
  internalWebPropertyId = NULL, name = NULL, parentLink = NULL,
  profileId = NULL, selfLink = NULL, type = NULL, updated = NULL,
  urlDestinationDetails = NULL, value = NULL, visitNumPagesDetails = NULL,
  visitTimeOnSiteDetails = NULL, webPropertyId = NULL)
}
\arguments{
\item{Goal.eventDetails}{The \link{Goal.eventDetails} object or list of objects}

\item{Goal.eventDetails.eventConditions}{The \link{Goal.eventDetails.eventConditions} object or list of objects}

\item{Goal.parentLink}{The \link{Goal.parentLink} object or list of objects}

\item{Goal.urlDestinationDetails}{The \link{Goal.urlDestinationDetails} object or list of objects}

\item{Goal.urlDestinationDetails.steps}{The \link{Goal.urlDestinationDetails.steps} object or list of objects}

\item{Goal.visitNumPagesDetails}{The \link{Goal.visitNumPagesDetails} object or list of objects}

\item{Goal.visitTimeOnSiteDetails}{The \link{Goal.visitTimeOnSiteDetails} object or list of objects}

\item{accountId}{Account ID to which this goal belongs}

\item{active}{Determines whether this goal is active}

\item{created}{Time this goal was created}

\item{eventDetails}{Details for the goal of the type EVENT}

\item{id}{Goal ID}

\item{internalWebPropertyId}{Internal ID for the web property to which this goal belongs}

\item{name}{Goal name}

\item{parentLink}{Parent link for a goal}

\item{profileId}{View (Profile) ID to which this goal belongs}

\item{selfLink}{Link for this goal}

\item{type}{Goal type}

\item{updated}{Time this goal was last modified}

\item{urlDestinationDetails}{Details for the goal of the type URL_DESTINATION}

\item{value}{Goal value}

\item{visitNumPagesDetails}{Details for the goal of the type VISIT_NUM_PAGES}

\item{visitTimeOnSiteDetails}{Details for the goal of the type VISIT_TIME_ON_SITE}

\item{webPropertyId}{Web property ID to which this goal belongs}
}
\value{
Goal object
}
\description{
Goal Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for Analytics goal resource.
}
\seealso{
Other Goal functions: \code{\link{Goal.eventDetails.eventConditions}},
  \code{\link{Goal.eventDetails}},
  \code{\link{Goal.parentLink}},
  \code{\link{Goal.urlDestinationDetails.steps}},
  \code{\link{Goal.urlDestinationDetails}},
  \code{\link{Goal.visitNumPagesDetails}},
  \code{\link{Goal.visitTimeOnSiteDetails}},
  \code{\link{management.goals.insert}},
  \code{\link{management.goals.patch}},
  \code{\link{management.goals.update}}
}

