% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Goal.urlDestinationDetails}
\alias{Goal.urlDestinationDetails}
\title{Goal.urlDestinationDetails Object}
\usage{
Goal.urlDestinationDetails(Goal.urlDestinationDetails.steps = NULL,
  caseSensitive = NULL, firstStepRequired = NULL, matchType = NULL,
  steps = NULL, url = NULL)
}
\arguments{
\item{Goal.urlDestinationDetails.steps}{The \link{Goal.urlDestinationDetails.steps} object or list of objects}

\item{caseSensitive}{Determines if the goal URL must exactly match the capitalization of visited URLs}

\item{firstStepRequired}{Determines if the first step in this goal is required}

\item{matchType}{Match type for the goal URL}

\item{steps}{List of steps configured for this goal funnel}

\item{url}{URL for this goal}
}
\value{
Goal.urlDestinationDetails object
}
\description{
Goal.urlDestinationDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details for the goal of the type URL_DESTINATION.
}
\seealso{
Other Goal functions: \code{\link{Goal.eventDetails.eventConditions}},
  \code{\link{Goal.eventDetails}},
  \code{\link{Goal.parentLink}},
  \code{\link{Goal.urlDestinationDetails.steps}},
  \code{\link{Goal.visitNumPagesDetails}},
  \code{\link{Goal.visitTimeOnSiteDetails}},
  \code{\link{Goal}},
  \code{\link{management.goals.insert}},
  \code{\link{management.goals.patch}},
  \code{\link{management.goals.update}}
}

