% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Goal.visitNumPagesDetails}
\alias{Goal.visitNumPagesDetails}
\title{Goal.visitNumPagesDetails Object}
\usage{
Goal.visitNumPagesDetails(comparisonType = NULL, comparisonValue = NULL)
}
\arguments{
\item{comparisonType}{Type of comparison}

\item{comparisonValue}{Value used for this comparison}
}
\value{
Goal.visitNumPagesDetails object
}
\description{
Goal.visitNumPagesDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details for the goal of the type VISIT_NUM_PAGES.
}
\seealso{
Other Goal functions: \code{\link{Goal.eventDetails.eventConditions}},
  \code{\link{Goal.eventDetails}},
  \code{\link{Goal.parentLink}},
  \code{\link{Goal.urlDestinationDetails.steps}},
  \code{\link{Goal.urlDestinationDetails}},
  \code{\link{Goal.visitTimeOnSiteDetails}},
  \code{\link{Goal}},
  \code{\link{management.goals.insert}},
  \code{\link{management.goals.patch}},
  \code{\link{management.goals.update}}
}
