% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Goal.visitTimeOnSiteDetails}
\alias{Goal.visitTimeOnSiteDetails}
\title{Goal.visitTimeOnSiteDetails Object}
\usage{
Goal.visitTimeOnSiteDetails(comparisonType = NULL, comparisonValue = NULL)
}
\arguments{
\item{comparisonType}{Type of comparison}

\item{comparisonValue}{Value used for this comparison}
}
\value{
Goal.visitTimeOnSiteDetails object
}
\description{
Goal.visitTimeOnSiteDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details for the goal of the type VISIT_TIME_ON_SITE.
}
\seealso{
Other Goal functions: \code{\link{Goal.eventDetails.eventConditions}},
  \code{\link{Goal.eventDetails}},
  \code{\link{Goal.parentLink}},
  \code{\link{Goal.urlDestinationDetails.steps}},
  \code{\link{Goal.urlDestinationDetails}},
  \code{\link{Goal.visitNumPagesDetails}},
  \code{\link{Goal}},
  \code{\link{management.goals.insert}},
  \code{\link{management.goals.patch}},
  \code{\link{management.goals.update}}
}
