% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{IncludeConditions}
\alias{IncludeConditions}
\title{IncludeConditions Object}
\usage{
IncludeConditions(daysToLookBack = NULL, isSmartList = NULL,
  membershipDurationDays = NULL, segment = NULL)
}
\arguments{
\item{daysToLookBack}{The look-back window lets you specify a time frame for evaluating the behavior that qualifies users for your audience}

\item{isSmartList}{Boolean indicating whether this segment is a smart list}

\item{membershipDurationDays}{Number of days (in the range 1 to 540) a user remains in the audience}

\item{segment}{The segment condition that will cause a user to be added to an audience}
}
\value{
IncludeConditions object
}
\description{
IncludeConditions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for an Analytics Remarketing Include Conditions.
}
