% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{LinkedForeignAccount}
\alias{LinkedForeignAccount}
\title{LinkedForeignAccount Object}
\usage{
LinkedForeignAccount(accountId = NULL, id = NULL, linkedAccountId = NULL,
  remarketingAudienceId = NULL, status = NULL, type = NULL,
  webPropertyId = NULL)
}
\arguments{
\item{accountId}{Account ID to which this linked foreign account belongs}

\item{id}{Entity ad account link ID}

\item{linkedAccountId}{The foreign account ID}

\item{remarketingAudienceId}{Remarketing audience ID to which this linked foreign account belongs}

\item{status}{The status of this foreign account link}

\item{type}{The type of the foreign account}

\item{webPropertyId}{Web property ID of the form UA-XXXXX-YY to which this linked foreign account belongs}
}
\value{
LinkedForeignAccount object
}
\description{
LinkedForeignAccount Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for an Analytics Remarketing Audience Foreign Link.
}

