% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.accountSummaries.list}
\alias{management.accountSummaries.list}
\title{Lists account summaries (lightweight tree comprised of accounts/properties/profiles) to which the user has access.}
\usage{
management.accountSummaries.list(max.results = NULL, start.index = NULL)
}
\arguments{
\item{max.results}{The maximum number of account summaries to include in this response, where the largest acceptable value is 1000}

\item{start.index}{An index of the first entity to retrieve}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics.edit
\item https://www.googleapis.com/auth/analytics.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics.edit, https://www.googleapis.com/auth/analytics.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}
}
