% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.accountUserLinks.list}
\alias{management.accountUserLinks.list}
\title{Lists account-user links for a given account.}
\usage{
management.accountUserLinks.list(accountId, max.results = NULL,
  start.index = NULL)
}
\arguments{
\item{accountId}{Account ID to retrieve the user links for}

\item{max.results}{The maximum number of account-user links to include in this response}

\item{start.index}{An index of the first account-user link to retrieve}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics.manage.users
\item https://www.googleapis.com/auth/analytics.manage.users.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics.manage.users, https://www.googleapis.com/auth/analytics.manage.users.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}
}

