% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.customDataSources.list}
\alias{management.customDataSources.list}
\title{List custom data sources to which the user has access.}
\usage{
management.customDataSources.list(accountId, webPropertyId,
  max.results = NULL, start.index = NULL)
}
\arguments{
\item{accountId}{Account Id for the custom data sources to retrieve}

\item{webPropertyId}{Web property Id for the custom data sources to retrieve}

\item{max.results}{The maximum number of custom data sources to include in this response}

\item{start.index}{A 1-based index of the first custom data source to retrieve}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics
\item https://www.googleapis.com/auth/analytics.edit
\item https://www.googleapis.com/auth/analytics.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics, https://www.googleapis.com/auth/analytics.edit, https://www.googleapis.com/auth/analytics.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}
}

