% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.customDimensions.insert}
\alias{management.customDimensions.insert}
\title{Create a new custom dimension.}
\usage{
management.customDimensions.insert(CustomDimension, accountId, webPropertyId)
}
\arguments{
\item{CustomDimension}{The \link{CustomDimension} object to pass to this method}

\item{accountId}{Account ID for the custom dimension to create}

\item{webPropertyId}{Web property ID for the custom dimension to create}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics.edit
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics.edit)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}

Other CustomDimension functions: \code{\link{CustomDimension.parentLink}},
  \code{\link{CustomDimension}},
  \code{\link{management.customDimensions.patch}},
  \code{\link{management.customDimensions.update}}
}
