% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.customMetrics.insert}
\alias{management.customMetrics.insert}
\title{Create a new custom metric.}
\usage{
management.customMetrics.insert(CustomMetric, accountId, webPropertyId)
}
\arguments{
\item{CustomMetric}{The \link{CustomMetric} object to pass to this method}

\item{accountId}{Account ID for the custom metric to create}

\item{webPropertyId}{Web property ID for the custom dimension to create}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics.edit
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics.edit)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}

Other CustomMetric functions: \code{\link{CustomMetric.parentLink}},
  \code{\link{CustomMetric}},
  \code{\link{management.customMetrics.patch}},
  \code{\link{management.customMetrics.update}}
}

