% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.customMetrics.update}
\alias{management.customMetrics.update}
\title{Updates an existing custom metric.}
\usage{
management.customMetrics.update(CustomMetric, accountId, webPropertyId,
  customMetricId, ignoreCustomDataSourceLinks = NULL)
}
\arguments{
\item{CustomMetric}{The \link{CustomMetric} object to pass to this method}

\item{accountId}{Account ID for the custom metric to update}

\item{webPropertyId}{Web property ID for the custom metric to update}

\item{customMetricId}{Custom metric ID for the custom metric to update}

\item{ignoreCustomDataSourceLinks}{Force the update and ignore any warnings related to the custom metric being linked to a custom data source / data set}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics.edit
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics.edit)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}

Other CustomMetric functions: \code{\link{CustomMetric.parentLink}},
  \code{\link{CustomMetric}},
  \code{\link{management.customMetrics.insert}},
  \code{\link{management.customMetrics.patch}}
}
