% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.experiments.delete}
\alias{management.experiments.delete}
\title{Delete an experiment.}
\usage{
management.experiments.delete(accountId, webPropertyId, profileId, experimentId)
}
\arguments{
\item{accountId}{Account ID to which the experiment belongs}

\item{webPropertyId}{Web property ID to which the experiment belongs}

\item{profileId}{View (Profile) ID to which the experiment belongs}

\item{experimentId}{ID of the experiment to delete}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics
\item https://www.googleapis.com/auth/analytics.edit
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics, https://www.googleapis.com/auth/analytics.edit)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}
}
