% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.experiments.patch}
\alias{management.experiments.patch}
\title{Update an existing experiment. This method supports patch semantics.}
\usage{
management.experiments.patch(Experiment, accountId, webPropertyId, profileId,
  experimentId)
}
\arguments{
\item{Experiment}{The \link{Experiment} object to pass to this method}

\item{accountId}{Account ID of the experiment to update}

\item{webPropertyId}{Web property ID of the experiment to update}

\item{profileId}{View (Profile) ID of the experiment to update}

\item{experimentId}{Experiment ID of the experiment to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics
\item https://www.googleapis.com/auth/analytics.edit
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics, https://www.googleapis.com/auth/analytics.edit)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}

Other Experiment functions: \code{\link{Experiment.parentLink}},
  \code{\link{Experiment.variations}},
  \code{\link{Experiment}},
  \code{\link{management.experiments.insert}},
  \code{\link{management.experiments.update}}
}
