% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.goals.get}
\alias{management.goals.get}
\title{Gets a goal to which the user has access.}
\usage{
management.goals.get(accountId, webPropertyId, profileId, goalId)
}
\arguments{
\item{accountId}{Account ID to retrieve the goal for}

\item{webPropertyId}{Web property ID to retrieve the goal for}

\item{profileId}{View (Profile) ID to retrieve the goal for}

\item{goalId}{Goal ID to retrieve the goal for}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics.edit
\item https://www.googleapis.com/auth/analytics.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics.edit, https://www.googleapis.com/auth/analytics.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}
}
