% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.goals.patch}
\alias{management.goals.patch}
\title{Updates an existing goal. This method supports patch semantics.}
\usage{
management.goals.patch(Goal, accountId, webPropertyId, profileId, goalId)
}
\arguments{
\item{Goal}{The \link{Goal} object to pass to this method}

\item{accountId}{Account ID to update the goal}

\item{webPropertyId}{Web property ID to update the goal}

\item{profileId}{View (Profile) ID to update the goal}

\item{goalId}{Index of the goal to be updated}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics.edit
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics.edit)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}

Other Goal functions: \code{\link{Goal.eventDetails.eventConditions}},
  \code{\link{Goal.eventDetails}},
  \code{\link{Goal.parentLink}},
  \code{\link{Goal.urlDestinationDetails.steps}},
  \code{\link{Goal.urlDestinationDetails}},
  \code{\link{Goal.visitNumPagesDetails}},
  \code{\link{Goal.visitTimeOnSiteDetails}},
  \code{\link{Goal}},
  \code{\link{management.goals.insert}},
  \code{\link{management.goals.update}}
}
