% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.profileFilterLinks.get}
\alias{management.profileFilterLinks.get}
\title{Returns a single profile filter link.}
\usage{
management.profileFilterLinks.get(accountId, webPropertyId, profileId, linkId)
}
\arguments{
\item{accountId}{Account ID to retrieve profile filter link for}

\item{webPropertyId}{Web property Id to retrieve profile filter link for}

\item{profileId}{Profile ID to retrieve filter link for}

\item{linkId}{ID of the profile filter link}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics.edit
\item https://www.googleapis.com/auth/analytics.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics.edit, https://www.googleapis.com/auth/analytics.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}
}
