% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.profileFilterLinks.insert}
\alias{management.profileFilterLinks.insert}
\title{Create a new profile filter link.}
\usage{
management.profileFilterLinks.insert(ProfileFilterLink, accountId,
  webPropertyId, profileId)
}
\arguments{
\item{ProfileFilterLink}{The \link{ProfileFilterLink} object to pass to this method}

\item{accountId}{Account ID to create profile filter link for}

\item{webPropertyId}{Web property Id to create profile filter link for}

\item{profileId}{Profile ID to create filter link for}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics.edit
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics.edit)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}

Other ProfileFilterLink functions: \code{\link{ProfileFilterLink}},
  \code{\link{management.profileFilterLinks.patch}},
  \code{\link{management.profileFilterLinks.update}}
}
