% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.profileFilterLinks.update}
\alias{management.profileFilterLinks.update}
\title{Update an existing profile filter link.}
\usage{
management.profileFilterLinks.update(ProfileFilterLink, accountId,
  webPropertyId, profileId, linkId)
}
\arguments{
\item{ProfileFilterLink}{The \link{ProfileFilterLink} object to pass to this method}

\item{accountId}{Account ID to which profile filter link belongs}

\item{webPropertyId}{Web property Id to which profile filter link belongs}

\item{profileId}{Profile ID to which filter link belongs}

\item{linkId}{ID of the profile filter link to be updated}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics.edit
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics.edit)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}

Other ProfileFilterLink functions: \code{\link{ProfileFilterLink}},
  \code{\link{management.profileFilterLinks.insert}},
  \code{\link{management.profileFilterLinks.patch}}
}
