% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.profiles.insert}
\alias{management.profiles.insert}
\title{Create a new view (profile).}
\usage{
management.profiles.insert(Profile, accountId, webPropertyId)
}
\arguments{
\item{Profile}{The \link{Profile} object to pass to this method}

\item{accountId}{Account ID to create the view (profile) for}

\item{webPropertyId}{Web property ID to create the view (profile) for}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics.edit
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics.edit)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}

Other Profile functions: \code{\link{Profile.childLink}},
  \code{\link{Profile.parentLink}},
  \code{\link{Profile.permissions}}, \code{\link{Profile}},
  \code{\link{management.profiles.patch}},
  \code{\link{management.profiles.update}}
}
