% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.profileUserLinks.list}
\alias{management.profileUserLinks.list}
\title{Lists profile-user links for a given view (profile).}
\usage{
management.profileUserLinks.list(accountId, webPropertyId, profileId,
  max.results = NULL, start.index = NULL)
}
\arguments{
\item{accountId}{Account ID which the given view (profile) belongs to}

\item{webPropertyId}{Web Property ID which the given view (profile) belongs to}

\item{profileId}{View (Profile) ID to retrieve the profile-user links for}

\item{max.results}{The maximum number of profile-user links to include in this response}

\item{start.index}{An index of the first profile-user link to retrieve}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics.manage.users
\item https://www.googleapis.com/auth/analytics.manage.users.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics.manage.users, https://www.googleapis.com/auth/analytics.manage.users.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}
}

