% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.remarketingAudience.list}
\alias{management.remarketingAudience.list}
\title{Lists remarketing audiences to which the user has access.}
\usage{
management.remarketingAudience.list(accountId, webPropertyId,
  max.results = NULL, start.index = NULL, type = NULL)
}
\arguments{
\item{accountId}{The account ID of the remarketing audiences to retrieve}

\item{webPropertyId}{The web property ID of the remarketing audiences to retrieve}

\item{max.results}{The maximum number of remarketing audiences to include in this response}

\item{start.index}{An index of the first entity to retrieve}

\item{type}{}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics.edit
\item https://www.googleapis.com/auth/analytics.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics.edit, https://www.googleapis.com/auth/analytics.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}
}
