% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.remarketingAudience.update}
\alias{management.remarketingAudience.update}
\title{Updates an existing remarketing audiences.}
\usage{
management.remarketingAudience.update(RemarketingAudience, accountId,
  webPropertyId, remarketingAudienceId)
}
\arguments{
\item{RemarketingAudience}{The \link{RemarketingAudience} object to pass to this method}

\item{accountId}{Account ID for the remarketing audience to update}

\item{webPropertyId}{Web property ID for the remarketing audience to update}

\item{remarketingAudienceId}{Remarketing audience ID of the remarketing audience to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics.edit
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics.edit)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}

Other RemarketingAudience functions: \code{\link{RemarketingAudience.audienceDefinition}},
  \code{\link{RemarketingAudience.stateBasedAudienceDefinition.excludeConditions}},
  \code{\link{RemarketingAudience.stateBasedAudienceDefinition}},
  \code{\link{RemarketingAudience}},
  \code{\link{management.remarketingAudience.insert}},
  \code{\link{management.remarketingAudience.patch}}
}

