% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.unsampledReports.insert}
\alias{management.unsampledReports.insert}
\title{Create a new unsampled report.}
\usage{
management.unsampledReports.insert(UnsampledReport, accountId, webPropertyId,
  profileId)
}
\arguments{
\item{UnsampledReport}{The \link{UnsampledReport} object to pass to this method}

\item{accountId}{Account ID to create the unsampled report for}

\item{webPropertyId}{Web property ID to create the unsampled report for}

\item{profileId}{View (Profile) ID to create the unsampled report for}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics
\item https://www.googleapis.com/auth/analytics.edit
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics, https://www.googleapis.com/auth/analytics.edit)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}

Other UnsampledReport functions: \code{\link{UnsampledReport.cloudStorageDownloadDetails}},
  \code{\link{UnsampledReport.driveDownloadDetails}},
  \code{\link{UnsampledReport}}
}

