% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.uploads.deleteUploadData}
\alias{management.uploads.deleteUploadData}
\title{Delete data associated with a previous upload.}
\usage{
management.uploads.deleteUploadData(AnalyticsDataimportDeleteUploadDataRequest,
  accountId, webPropertyId, customDataSourceId)
}
\arguments{
\item{AnalyticsDataimportDeleteUploadDataRequest}{The \link{AnalyticsDataimportDeleteUploadDataRequest} object to pass to this method}

\item{accountId}{Account Id for the uploads to be deleted}

\item{webPropertyId}{Web property Id for the uploads to be deleted}

\item{customDataSourceId}{Custom data source Id for the uploads to be deleted}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics
\item https://www.googleapis.com/auth/analytics.edit
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics, https://www.googleapis.com/auth/analytics.edit)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}

Other AnalyticsDataimportDeleteUploadDataRequest functions: \code{\link{AnalyticsDataimportDeleteUploadDataRequest}}
}

