% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.uploads.get}
\alias{management.uploads.get}
\title{List uploads to which the user has access.}
\usage{
management.uploads.get(accountId, webPropertyId, customDataSourceId, uploadId)
}
\arguments{
\item{accountId}{Account Id for the upload to retrieve}

\item{webPropertyId}{Web property Id for the upload to retrieve}

\item{customDataSourceId}{Custom data source Id for upload to retrieve}

\item{uploadId}{Upload Id to retrieve}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics
\item https://www.googleapis.com/auth/analytics.edit
\item https://www.googleapis.com/auth/analytics.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics, https://www.googleapis.com/auth/analytics.edit, https://www.googleapis.com/auth/analytics.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}
}

