% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.uploads.uploadData}
\alias{management.uploads.uploadData}
\title{Upload data for a custom data source.}
\usage{
management.uploads.uploadData(accountId, webPropertyId, customDataSourceId)
}
\arguments{
\item{accountId}{Account Id associated with the upload}

\item{webPropertyId}{Web property UA-string associated with the upload}

\item{customDataSourceId}{Custom data source Id to which the data being uploaded belongs}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics
\item https://www.googleapis.com/auth/analytics.edit
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics, https://www.googleapis.com/auth/analytics.edit)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}
}
