% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.webproperties.patch}
\alias{management.webproperties.patch}
\title{Updates an existing web property. This method supports patch semantics.}
\usage{
management.webproperties.patch(Webproperty, accountId, webPropertyId)
}
\arguments{
\item{Webproperty}{The \link{Webproperty} object to pass to this method}

\item{accountId}{Account ID to which the web property belongs}

\item{webPropertyId}{Web property ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics.edit
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics.edit)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}

Other Webproperty functions: \code{\link{Webproperty.childLink}},
  \code{\link{Webproperty.parentLink}},
  \code{\link{Webproperty.permissions}},
  \code{\link{Webproperty}},
  \code{\link{management.webproperties.insert}},
  \code{\link{management.webproperties.update}}
}
