% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.webPropertyAdWordsLinks.list}
\alias{management.webPropertyAdWordsLinks.list}
\title{Lists webProperty-AdWords links for a given web property.}
\usage{
management.webPropertyAdWordsLinks.list(accountId, webPropertyId,
  max.results = NULL, start.index = NULL)
}
\arguments{
\item{accountId}{ID of the account which the given web property belongs to}

\item{webPropertyId}{Web property ID to retrieve the AdWords links for}

\item{max.results}{The maximum number of webProperty-AdWords links to include in this response}

\item{start.index}{An index of the first webProperty-AdWords link to retrieve}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics.edit
\item https://www.googleapis.com/auth/analytics.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics.edit, https://www.googleapis.com/auth/analytics.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}
}
