% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.webpropertyUserLinks.list}
\alias{management.webpropertyUserLinks.list}
\title{Lists webProperty-user links for a given web property.}
\usage{
management.webpropertyUserLinks.list(accountId, webPropertyId,
  max.results = NULL, start.index = NULL)
}
\arguments{
\item{accountId}{Account ID which the given web property belongs to}

\item{webPropertyId}{Web Property ID for the webProperty-user links to retrieve}

\item{max.results}{The maximum number of webProperty-user Links to include in this response}

\item{start.index}{An index of the first webProperty-user link to retrieve}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics.manage.users
\item https://www.googleapis.com/auth/analytics.manage.users.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics.manage.users, https://www.googleapis.com/auth/analytics.manage.users.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}
}
