% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.webpropertyUserLinks.update}
\alias{management.webpropertyUserLinks.update}
\title{Updates permissions for an existing user on the given web property.}
\usage{
management.webpropertyUserLinks.update(EntityUserLink, accountId, webPropertyId,
  linkId)
}
\arguments{
\item{EntityUserLink}{The \link{EntityUserLink} object to pass to this method}

\item{accountId}{Account ID to update the account-user link for}

\item{webPropertyId}{Web property ID to update the account-user link for}

\item{linkId}{Link ID to update the account-user link for}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics.manage.users
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics.manage.users)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}

Other EntityUserLink functions: \code{\link{EntityUserLink.entity}},
  \code{\link{EntityUserLink.permissions}},
  \code{\link{EntityUserLink}},
  \code{\link{management.accountUserLinks.insert}},
  \code{\link{management.accountUserLinks.update}},
  \code{\link{management.profileUserLinks.insert}},
  \code{\link{management.profileUserLinks.update}},
  \code{\link{management.webpropertyUserLinks.insert}}
}
