% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{McfData.columnHeaders}
\alias{McfData.columnHeaders}
\title{McfData.columnHeaders Object}
\usage{
McfData.columnHeaders()
}
\value{
McfData.columnHeaders object
}
\description{
McfData.columnHeaders Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other McfData functions: \code{\link{McfData.profileInfo}},
  \code{\link{McfData.query}},
  \code{\link{McfData.totalsForAllResults}},
  \code{\link{McfData}}
}

