% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{McfData.profileInfo}
\alias{McfData.profileInfo}
\title{McfData.profileInfo Object}
\usage{
McfData.profileInfo(accountId = NULL, internalWebPropertyId = NULL,
  profileId = NULL, profileName = NULL, tableId = NULL,
  webPropertyId = NULL)
}
\arguments{
\item{accountId}{Account ID to which this view (profile) belongs}

\item{internalWebPropertyId}{Internal ID for the web property to which this view (profile) belongs}

\item{profileId}{View (Profile) ID}

\item{profileName}{View (Profile) name}

\item{tableId}{Table ID for view (profile)}

\item{webPropertyId}{Web Property ID to which this view (profile) belongs}
}
\value{
McfData.profileInfo object
}
\description{
McfData.profileInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information for the view (profile), for which the Analytics data was requested.
}
\seealso{
Other McfData functions: \code{\link{McfData.columnHeaders}},
  \code{\link{McfData.query}},
  \code{\link{McfData.totalsForAllResults}},
  \code{\link{McfData}}
}
