% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{McfData}
\alias{McfData}
\title{McfData Object}
\usage{
McfData(McfData.columnHeaders = NULL, McfData.profileInfo = NULL,
  McfData.query = NULL, McfData.totalsForAllResults = NULL,
  columnHeaders = NULL, containsSampledData = NULL, id = NULL,
  itemsPerPage = NULL, nextLink = NULL, previousLink = NULL,
  profileInfo = NULL, query = NULL, rows = NULL, sampleSize = NULL,
  sampleSpace = NULL, selfLink = NULL, totalResults = NULL,
  totalsForAllResults = NULL)
}
\arguments{
\item{McfData.columnHeaders}{The \link{McfData.columnHeaders} object or list of objects}

\item{McfData.profileInfo}{The \link{McfData.profileInfo} object or list of objects}

\item{McfData.query}{The \link{McfData.query} object or list of objects}

\item{McfData.totalsForAllResults}{The \link{McfData.totalsForAllResults} object or list of objects}

\item{columnHeaders}{Column headers that list dimension names followed by the metric names}

\item{containsSampledData}{Determines if the Analytics data contains sampled data}

\item{id}{Unique ID for this data response}

\item{itemsPerPage}{The maximum number of rows the response can contain, regardless of the actual number of rows returned}

\item{nextLink}{Link to next page for this Analytics data query}

\item{previousLink}{Link to previous page for this Analytics data query}

\item{profileInfo}{Information for the view (profile), for which the Analytics data was requested}

\item{query}{Analytics data request query parameters}

\item{rows}{Analytics data rows, where each row contains a list of dimension values followed by the metric values}

\item{sampleSize}{The number of samples used to calculate the result}

\item{sampleSpace}{Total size of the sample space from which the samples were selected}

\item{selfLink}{Link to this page}

\item{totalResults}{The total number of rows for the query, regardless of the number of rows in the response}

\item{totalsForAllResults}{Total values for the requested metrics over all the results, not just the results returned in this response}
}
\value{
McfData object
}
\description{
McfData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Multi-Channel Funnels data for a given view (profile).
}
\seealso{
Other McfData functions: \code{\link{McfData.columnHeaders}},
  \code{\link{McfData.profileInfo}},
  \code{\link{McfData.query}},
  \code{\link{McfData.totalsForAllResults}}
}
