% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Profile.childLink}
\alias{Profile.childLink}
\title{Profile.childLink Object}
\usage{
Profile.childLink(href = NULL, type = NULL)
}
\arguments{
\item{href}{Link to the list of goals for this view (profile)}

\item{type}{Value is 'analytics#goals'}
}
\value{
Profile.childLink object
}
\description{
Profile.childLink Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Child link for this view (profile). Points to the list of goals for this view (profile).
}
\seealso{
Other Profile functions: \code{\link{Profile.parentLink}},
  \code{\link{Profile.permissions}}, \code{\link{Profile}},
  \code{\link{management.profiles.insert}},
  \code{\link{management.profiles.patch}},
  \code{\link{management.profiles.update}}
}

