% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Profile.parentLink}
\alias{Profile.parentLink}
\title{Profile.parentLink Object}
\usage{
Profile.parentLink(href = NULL, type = NULL)
}
\arguments{
\item{href}{Link to the web property to which this view (profile) belongs}

\item{type}{Value is 'analytics#webproperty'}
}
\value{
Profile.parentLink object
}
\description{
Profile.parentLink Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Parent link for this view (profile). Points to the web property to which this view (profile) belongs.
}
\seealso{
Other Profile functions: \code{\link{Profile.childLink}},
  \code{\link{Profile.permissions}}, \code{\link{Profile}},
  \code{\link{management.profiles.insert}},
  \code{\link{management.profiles.patch}},
  \code{\link{management.profiles.update}}
}
