% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Profile.permissions}
\alias{Profile.permissions}
\title{Profile.permissions Object}
\usage{
Profile.permissions()
}
\value{
Profile.permissions object
}
\description{
Profile.permissions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Permissions the user has for this view (profile).
}
\seealso{
Other Profile functions: \code{\link{Profile.childLink}},
  \code{\link{Profile.parentLink}}, \code{\link{Profile}},
  \code{\link{management.profiles.insert}},
  \code{\link{management.profiles.patch}},
  \code{\link{management.profiles.update}}
}

