% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Profile}
\alias{Profile}
\title{Profile Object}
\usage{
Profile(Profile.childLink = NULL, Profile.parentLink = NULL,
  Profile.permissions = NULL, accountId = NULL,
  botFilteringEnabled = NULL, childLink = NULL, currency = NULL,
  defaultPage = NULL, eCommerceTracking = NULL,
  enhancedECommerceTracking = NULL, excludeQueryParameters = NULL,
  id = NULL, name = NULL, parentLink = NULL, permissions = NULL,
  siteSearchCategoryParameters = NULL, siteSearchQueryParameters = NULL,
  starred = NULL, stripSiteSearchCategoryParameters = NULL,
  stripSiteSearchQueryParameters = NULL, timezone = NULL, type = NULL,
  websiteUrl = NULL)
}
\arguments{
\item{Profile.childLink}{The \link{Profile.childLink} object or list of objects}

\item{Profile.parentLink}{The \link{Profile.parentLink} object or list of objects}

\item{Profile.permissions}{The \link{Profile.permissions} object or list of objects}

\item{accountId}{Account ID to which this view (profile) belongs}

\item{botFilteringEnabled}{Indicates whether bot filtering is enabled for this view (profile)}

\item{childLink}{Child link for this view (profile)}

\item{currency}{The currency type associated with this view (profile), defaults to USD}

\item{defaultPage}{Default page for this view (profile)}

\item{eCommerceTracking}{Indicates whether ecommerce tracking is enabled for this view (profile)}

\item{enhancedECommerceTracking}{Indicates whether enhanced ecommerce tracking is enabled for this view (profile)}

\item{excludeQueryParameters}{The query parameters that are excluded from this view (profile)}

\item{id}{View (Profile) ID}

\item{name}{Name of this view (profile)}

\item{parentLink}{Parent link for this view (profile)}

\item{permissions}{Permissions the user has for this view (profile)}

\item{siteSearchCategoryParameters}{Site search category parameters for this view (profile)}

\item{siteSearchQueryParameters}{The site search query parameters for this view (profile)}

\item{starred}{Indicates whether this view (profile) is starred or not}

\item{stripSiteSearchCategoryParameters}{Whether or not Analytics will strip search category parameters from the URLs in your reports}

\item{stripSiteSearchQueryParameters}{Whether or not Analytics will strip search query parameters from the URLs in your reports}

\item{timezone}{Time zone for which this view (profile) has been configured}

\item{type}{View (Profile) type}

\item{websiteUrl}{Website URL for this view (profile)}
}
\value{
Profile object
}
\description{
Profile Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for an Analytics view (profile).
}
\seealso{
Other Profile functions: \code{\link{Profile.childLink}},
  \code{\link{Profile.parentLink}},
  \code{\link{Profile.permissions}},
  \code{\link{management.profiles.insert}},
  \code{\link{management.profiles.patch}},
  \code{\link{management.profiles.update}}
}

