% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{ProfileFilterLink}
\alias{ProfileFilterLink}
\title{ProfileFilterLink Object}
\usage{
ProfileFilterLink(filterRef = NULL, id = NULL, profileRef = NULL,
  rank = NULL)
}
\arguments{
\item{filterRef}{Filter for this link}

\item{id}{Profile filter link ID}

\item{profileRef}{View (Profile) for this link}

\item{rank}{The rank of this profile filter link relative to the other filters linked to the same profile}
}
\value{
ProfileFilterLink object
}
\description{
ProfileFilterLink Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for an Analytics profile filter link.
}
\seealso{
Other ProfileFilterLink functions: \code{\link{management.profileFilterLinks.insert}},
  \code{\link{management.profileFilterLinks.patch}},
  \code{\link{management.profileFilterLinks.update}}
}

