% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{ProfileRef}
\alias{ProfileRef}
\title{ProfileRef Object}
\usage{
ProfileRef(accountId = NULL, href = NULL, id = NULL,
  internalWebPropertyId = NULL, name = NULL, webPropertyId = NULL)
}
\arguments{
\item{accountId}{Account ID to which this view (profile) belongs}

\item{href}{Link for this view (profile)}

\item{id}{View (Profile) ID}

\item{internalWebPropertyId}{Internal ID for the web property to which this view (profile) belongs}

\item{name}{Name of this view (profile)}

\item{webPropertyId}{Web property ID of the form UA-XXXXX-YY to which this view (profile) belongs}
}
\value{
ProfileRef object
}
\description{
ProfileRef Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for a linked view (profile).
}

