% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{ProfileSummary}
\alias{ProfileSummary}
\title{ProfileSummary Object}
\usage{
ProfileSummary(id = NULL, name = NULL, starred = NULL, type = NULL)
}
\arguments{
\item{id}{View (profile) ID}

\item{name}{View (profile) name}

\item{starred}{Indicates whether this view (profile) is starred or not}

\item{type}{View (Profile) type}
}
\value{
ProfileSummary object
}
\description{
ProfileSummary Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for an Analytics ProfileSummary. ProfileSummary returns basic information (i.e., summary) for a profile.
}

