% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{provisioning.createAccountTicket}
\alias{provisioning.createAccountTicket}
\title{Creates an account ticket.}
\usage{
provisioning.createAccountTicket(AccountTicket)
}
\arguments{
\item{AccountTicket}{The \link{AccountTicket} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics.provision
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics.provision)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}

Other AccountTicket functions: \code{\link{AccountTicket}}
}
