% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{RealtimeData.columnHeaders}
\alias{RealtimeData.columnHeaders}
\title{RealtimeData.columnHeaders Object}
\usage{
RealtimeData.columnHeaders()
}
\value{
RealtimeData.columnHeaders object
}
\description{
RealtimeData.columnHeaders Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other RealtimeData functions: \code{\link{RealtimeData.profileInfo}},
  \code{\link{RealtimeData.query}},
  \code{\link{RealtimeData.totalsForAllResults}},
  \code{\link{RealtimeData}}
}

