% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{RealtimeData.profileInfo}
\alias{RealtimeData.profileInfo}
\title{RealtimeData.profileInfo Object}
\usage{
RealtimeData.profileInfo(accountId = NULL, internalWebPropertyId = NULL,
  profileId = NULL, profileName = NULL, tableId = NULL,
  webPropertyId = NULL)
}
\arguments{
\item{accountId}{Account ID to which this view (profile) belongs}

\item{internalWebPropertyId}{Internal ID for the web property to which this view (profile) belongs}

\item{profileId}{View (Profile) ID}

\item{profileName}{View (Profile) name}

\item{tableId}{Table ID for view (profile)}

\item{webPropertyId}{Web Property ID to which this view (profile) belongs}
}
\value{
RealtimeData.profileInfo object
}
\description{
RealtimeData.profileInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information for the view (profile), for which the real time data was requested.
}
\seealso{
Other RealtimeData functions: \code{\link{RealtimeData.columnHeaders}},
  \code{\link{RealtimeData.query}},
  \code{\link{RealtimeData.totalsForAllResults}},
  \code{\link{RealtimeData}}
}
