% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{RealtimeData.query}
\alias{RealtimeData.query}
\title{RealtimeData.query Object}
\usage{
RealtimeData.query(dimensions = NULL, filters = NULL, ids = NULL,
  max.results = NULL, metrics = NULL, sort = NULL)
}
\arguments{
\item{dimensions}{List of real time dimensions}

\item{filters}{Comma-separated list of dimension or metric filters}

\item{ids}{Unique table ID}

\item{max.results}{Maximum results per page}

\item{metrics}{List of real time metrics}

\item{sort}{List of dimensions or metrics based on which real time data is sorted}
}
\value{
RealtimeData.query object
}
\description{
RealtimeData.query Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Real time data request query parameters.
}
\seealso{
Other RealtimeData functions: \code{\link{RealtimeData.columnHeaders}},
  \code{\link{RealtimeData.profileInfo}},
  \code{\link{RealtimeData.totalsForAllResults}},
  \code{\link{RealtimeData}}
}
