% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{RemarketingAudience.audienceDefinition}
\alias{RemarketingAudience.audienceDefinition}
\title{RemarketingAudience.audienceDefinition Object}
\usage{
RemarketingAudience.audienceDefinition(includeConditions = NULL)
}
\arguments{
\item{includeConditions}{Defines the conditions to include users to the audience}
}
\value{
RemarketingAudience.audienceDefinition object
}
\description{
RemarketingAudience.audienceDefinition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The simple audience definition that will cause a user to be added to an audience.
}
\seealso{
Other RemarketingAudience functions: \code{\link{RemarketingAudience.stateBasedAudienceDefinition.excludeConditions}},
  \code{\link{RemarketingAudience.stateBasedAudienceDefinition}},
  \code{\link{RemarketingAudience}},
  \code{\link{management.remarketingAudience.insert}},
  \code{\link{management.remarketingAudience.patch}},
  \code{\link{management.remarketingAudience.update}}
}

