% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{RemarketingAudience}
\alias{RemarketingAudience}
\title{RemarketingAudience Object}
\usage{
RemarketingAudience(RemarketingAudience.audienceDefinition = NULL,
  RemarketingAudience.stateBasedAudienceDefinition = NULL,
  RemarketingAudience.stateBasedAudienceDefinition.excludeConditions = NULL,
  accountId = NULL, audienceDefinition = NULL, audienceType = NULL,
  id = NULL, linkedAdAccounts = NULL, linkedViews = NULL, name = NULL,
  stateBasedAudienceDefinition = NULL, webPropertyId = NULL)
}
\arguments{
\item{RemarketingAudience.audienceDefinition}{The \link{RemarketingAudience.audienceDefinition} object or list of objects}

\item{RemarketingAudience.stateBasedAudienceDefinition}{The \link{RemarketingAudience.stateBasedAudienceDefinition} object or list of objects}

\item{RemarketingAudience.stateBasedAudienceDefinition.excludeConditions}{The \link{RemarketingAudience.stateBasedAudienceDefinition.excludeConditions} object or list of objects}

\item{accountId}{Account ID to which this remarketing audience belongs}

\item{audienceDefinition}{The simple audience definition that will cause a user to be added to an audience}

\item{audienceType}{The type of audience, either SIMPLE or STATE_BASED}

\item{id}{Remarketing Audience ID}

\item{linkedAdAccounts}{The linked ad accounts associated with this remarketing audience}

\item{linkedViews}{The views (profiles) that this remarketing audience is linked to}

\item{name}{The name of this remarketing audience}

\item{stateBasedAudienceDefinition}{A state based audience definition that will cause a user to be added or removed from an audience}

\item{webPropertyId}{Web property ID of the form UA-XXXXX-YY to which this remarketing audience belongs}
}
\value{
RemarketingAudience object
}
\description{
RemarketingAudience Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for an Analytics remarketing audience.
}
\seealso{
Other RemarketingAudience functions: \code{\link{RemarketingAudience.audienceDefinition}},
  \code{\link{RemarketingAudience.stateBasedAudienceDefinition.excludeConditions}},
  \code{\link{RemarketingAudience.stateBasedAudienceDefinition}},
  \code{\link{management.remarketingAudience.insert}},
  \code{\link{management.remarketingAudience.patch}},
  \code{\link{management.remarketingAudience.update}}
}
