% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{RemarketingAudience.stateBasedAudienceDefinition.excludeConditions}
\alias{RemarketingAudience.stateBasedAudienceDefinition.excludeConditions}
\title{RemarketingAudience.stateBasedAudienceDefinition.excludeConditions Object}
\usage{

  RemarketingAudience.stateBasedAudienceDefinition.excludeConditions(exclusionDuration = NULL,
  segment = NULL)
}
\arguments{
\item{exclusionDuration}{Whether to make the exclusion TEMPORARY or PERMANENT}

\item{segment}{The segment condition that will cause a user to be removed from an audience}
}
\value{
RemarketingAudience.stateBasedAudienceDefinition.excludeConditions object
}
\description{
RemarketingAudience.stateBasedAudienceDefinition.excludeConditions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Defines the conditions to exclude users from the audience.
}
\seealso{
Other RemarketingAudience functions: \code{\link{RemarketingAudience.audienceDefinition}},
  \code{\link{RemarketingAudience.stateBasedAudienceDefinition}},
  \code{\link{RemarketingAudience}},
  \code{\link{management.remarketingAudience.insert}},
  \code{\link{management.remarketingAudience.patch}},
  \code{\link{management.remarketingAudience.update}}
}
