% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{RemarketingAudience.stateBasedAudienceDefinition}
\alias{RemarketingAudience.stateBasedAudienceDefinition}
\title{RemarketingAudience.stateBasedAudienceDefinition Object}
\usage{

  RemarketingAudience.stateBasedAudienceDefinition(RemarketingAudience.stateBasedAudienceDefinition.excludeConditions = NULL,
  excludeConditions = NULL, includeConditions = NULL)
}
\arguments{
\item{RemarketingAudience.stateBasedAudienceDefinition.excludeConditions}{The \link{RemarketingAudience.stateBasedAudienceDefinition.excludeConditions} object or list of objects}

\item{excludeConditions}{Defines the conditions to exclude users from the audience}

\item{includeConditions}{Defines the conditions to include users to the audience}
}
\value{
RemarketingAudience.stateBasedAudienceDefinition object
}
\description{
RemarketingAudience.stateBasedAudienceDefinition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A state based audience definition that will cause a user to be added or removed from an audience.
}
\seealso{
Other RemarketingAudience functions: \code{\link{RemarketingAudience.audienceDefinition}},
  \code{\link{RemarketingAudience.stateBasedAudienceDefinition.excludeConditions}},
  \code{\link{RemarketingAudience}},
  \code{\link{management.remarketingAudience.insert}},
  \code{\link{management.remarketingAudience.patch}},
  \code{\link{management.remarketingAudience.update}}
}

