% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{RemarketingAudiences}
\alias{RemarketingAudiences}
\title{RemarketingAudiences Object}
\usage{
RemarketingAudiences(items = NULL, itemsPerPage = NULL, nextLink = NULL,
  previousLink = NULL, startIndex = NULL, totalResults = NULL,
  username = NULL)
}
\arguments{
\item{items}{A list of remarketing audiences}

\item{itemsPerPage}{The maximum number of resources the response can contain, regardless of the actual number of resources returned}

\item{nextLink}{Link to next page for this remarketing audience collection}

\item{previousLink}{Link to previous page for this view (profile) collection}

\item{startIndex}{The starting index of the resources, which is 1 by default or otherwise specified by the start-index query parameter}

\item{totalResults}{The total number of results for the query, regardless of the number of results in the response}

\item{username}{Email ID of the authenticated user}
}
\value{
RemarketingAudiences object
}
\description{
RemarketingAudiences Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A remarketing audience collection lists Analytics remarketing audiences to which the user has access. Each resource in the collection corresponds to a single Analytics remarketing audience.
}

