% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Segment}
\alias{Segment}
\title{Segment Object}
\usage{
Segment(created = NULL, definition = NULL, id = NULL, name = NULL,
  segmentId = NULL, selfLink = NULL, type = NULL, updated = NULL)
}
\arguments{
\item{created}{Time the segment was created}

\item{definition}{Segment definition}

\item{id}{Segment ID}

\item{name}{Segment name}

\item{segmentId}{Segment ID}

\item{selfLink}{Link for this segment}

\item{type}{Type for a segment}

\item{updated}{Time the segment was last modified}
}
\value{
Segment object
}
\description{
Segment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for an Analytics segment.
}

