% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{UnsampledReport.cloudStorageDownloadDetails}
\alias{UnsampledReport.cloudStorageDownloadDetails}
\title{UnsampledReport.cloudStorageDownloadDetails Object}
\usage{
UnsampledReport.cloudStorageDownloadDetails(bucketId = NULL,
  objectId = NULL)
}
\arguments{
\item{bucketId}{Id of the bucket the file object is stored in}

\item{objectId}{Id of the file object containing the report data}
}
\value{
UnsampledReport.cloudStorageDownloadDetails object
}
\description{
UnsampledReport.cloudStorageDownloadDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Download details for a file stored in Google Cloud Storage.
}
\seealso{
Other UnsampledReport functions: \code{\link{UnsampledReport.driveDownloadDetails}},
  \code{\link{UnsampledReport}},
  \code{\link{management.unsampledReports.insert}}
}
