% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{UnsampledReport.driveDownloadDetails}
\alias{UnsampledReport.driveDownloadDetails}
\title{UnsampledReport.driveDownloadDetails Object}
\usage{
UnsampledReport.driveDownloadDetails(documentId = NULL)
}
\arguments{
\item{documentId}{Id of the document/file containing the report data}
}
\value{
UnsampledReport.driveDownloadDetails object
}
\description{
UnsampledReport.driveDownloadDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Download details for a file stored in Google Drive.
}
\seealso{
Other UnsampledReport functions: \code{\link{UnsampledReport.cloudStorageDownloadDetails}},
  \code{\link{UnsampledReport}},
  \code{\link{management.unsampledReports.insert}}
}
