% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{UnsampledReport}
\alias{UnsampledReport}
\title{UnsampledReport Object}
\usage{
UnsampledReport(UnsampledReport.cloudStorageDownloadDetails = NULL,
  UnsampledReport.driveDownloadDetails = NULL, accountId = NULL,
  dimensions = NULL, end.date = NULL, filters = NULL, id = NULL,
  metrics = NULL, profileId = NULL, segment = NULL, start.date = NULL,
  title = NULL, webPropertyId = NULL)
}
\arguments{
\item{UnsampledReport.cloudStorageDownloadDetails}{The \link{UnsampledReport.cloudStorageDownloadDetails} object or list of objects}

\item{UnsampledReport.driveDownloadDetails}{The \link{UnsampledReport.driveDownloadDetails} object or list of objects}

\item{accountId}{Account ID to which this unsampled report belongs}

\item{dimensions}{The dimensions for the unsampled report}

\item{end.date}{The end date for the unsampled report}

\item{filters}{The filters for the unsampled report}

\item{id}{Unsampled report ID}

\item{metrics}{The metrics for the unsampled report}

\item{profileId}{View (Profile) ID to which this unsampled report belongs}

\item{segment}{The segment for the unsampled report}

\item{start.date}{The start date for the unsampled report}

\item{title}{Title of the unsampled report}

\item{webPropertyId}{Web property ID to which this unsampled report belongs}
}
\value{
UnsampledReport object
}
\description{
UnsampledReport Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for Analytics unsampled report resource.
}
\seealso{
Other UnsampledReport functions: \code{\link{UnsampledReport.cloudStorageDownloadDetails}},
  \code{\link{UnsampledReport.driveDownloadDetails}},
  \code{\link{management.unsampledReports.insert}}
}
