% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Upload}
\alias{Upload}
\title{Upload Object}
\usage{
Upload(accountId = NULL, customDataSourceId = NULL, errors = NULL,
  id = NULL, status = NULL)
}
\arguments{
\item{accountId}{Account Id to which this upload belongs}

\item{customDataSourceId}{Custom data source Id to which this data import belongs}

\item{errors}{Data import errors collection}

\item{id}{A unique ID for this upload}

\item{status}{Upload status}
}
\value{
Upload object
}
\description{
Upload Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata returned for an upload operation.
}
