% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Uploads}
\alias{Uploads}
\title{Uploads Object}
\usage{
Uploads(items = NULL, itemsPerPage = NULL, nextLink = NULL,
  previousLink = NULL, startIndex = NULL, totalResults = NULL)
}
\arguments{
\item{items}{A list of uploads}

\item{itemsPerPage}{The maximum number of resources the response can contain, regardless of the actual number of resources returned}

\item{nextLink}{Link to next page for this upload collection}

\item{previousLink}{Link to previous page for this upload collection}

\item{startIndex}{The starting index of the resources, which is 1 by default or otherwise specified by the start-index query parameter}

\item{totalResults}{The total number of results for the query, regardless of the number of resources in the result}
}
\value{
Uploads object
}
\description{
Uploads Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Upload collection lists Analytics uploads to which the user has access. Each custom data source can have a set of uploads. Each resource in the upload collection corresponds to a single Analytics data upload.
}

