% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{UserRef}
\alias{UserRef}
\title{UserRef Object}
\usage{
UserRef(email = NULL, id = NULL)
}
\arguments{
\item{email}{Email ID of this user}

\item{id}{User ID}
}
\value{
UserRef object
}
\description{
UserRef Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for a user reference.
}
