% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Webproperty.childLink}
\alias{Webproperty.childLink}
\title{Webproperty.childLink Object}
\usage{
Webproperty.childLink(href = NULL, type = NULL)
}
\arguments{
\item{href}{Link to the list of views (profiles) for this web property}

\item{type}{Type of the parent link}
}
\value{
Webproperty.childLink object
}
\description{
Webproperty.childLink Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Child link for this web property. Points to the list of views (profiles) for this web property.
}
\seealso{
Other Webproperty functions: \code{\link{Webproperty.parentLink}},
  \code{\link{Webproperty.permissions}},
  \code{\link{Webproperty}},
  \code{\link{management.webproperties.insert}},
  \code{\link{management.webproperties.patch}},
  \code{\link{management.webproperties.update}}
}

