% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Webproperty.parentLink}
\alias{Webproperty.parentLink}
\title{Webproperty.parentLink Object}
\usage{
Webproperty.parentLink(href = NULL, type = NULL)
}
\arguments{
\item{href}{Link to the account for this web property}

\item{type}{Type of the parent link}
}
\value{
Webproperty.parentLink object
}
\description{
Webproperty.parentLink Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Parent link for this web property. Points to the account to which this web property belongs.
}
\seealso{
Other Webproperty functions: \code{\link{Webproperty.childLink}},
  \code{\link{Webproperty.permissions}},
  \code{\link{Webproperty}},
  \code{\link{management.webproperties.insert}},
  \code{\link{management.webproperties.patch}},
  \code{\link{management.webproperties.update}}
}
