% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Webproperty.permissions}
\alias{Webproperty.permissions}
\title{Webproperty.permissions Object}
\usage{
Webproperty.permissions()
}
\value{
Webproperty.permissions object
}
\description{
Webproperty.permissions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Permissions the user has for this web property.
}
\seealso{
Other Webproperty functions: \code{\link{Webproperty.childLink}},
  \code{\link{Webproperty.parentLink}},
  \code{\link{Webproperty}},
  \code{\link{management.webproperties.insert}},
  \code{\link{management.webproperties.patch}},
  \code{\link{management.webproperties.update}}
}

