% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{Webproperty}
\alias{Webproperty}
\title{Webproperty Object}
\usage{
Webproperty(Webproperty.childLink = NULL, Webproperty.parentLink = NULL,
  Webproperty.permissions = NULL, accountId = NULL, childLink = NULL,
  defaultProfileId = NULL, id = NULL, industryVertical = NULL,
  name = NULL, parentLink = NULL, permissions = NULL, starred = NULL,
  websiteUrl = NULL)
}
\arguments{
\item{Webproperty.childLink}{The \link{Webproperty.childLink} object or list of objects}

\item{Webproperty.parentLink}{The \link{Webproperty.parentLink} object or list of objects}

\item{Webproperty.permissions}{The \link{Webproperty.permissions} object or list of objects}

\item{accountId}{Account ID to which this web property belongs}

\item{childLink}{Child link for this web property}

\item{defaultProfileId}{Default view (profile) ID}

\item{id}{Web property ID of the form UA-XXXXX-YY}

\item{industryVertical}{The industry vertical/category selected for this web property}

\item{name}{Name of this web property}

\item{parentLink}{Parent link for this web property}

\item{permissions}{Permissions the user has for this web property}

\item{starred}{Indicates whether this web property is starred or not}

\item{websiteUrl}{Website url for this web property}
}
\value{
Webproperty object
}
\description{
Webproperty Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for an Analytics web property.
}
\seealso{
Other Webproperty functions: \code{\link{Webproperty.childLink}},
  \code{\link{Webproperty.parentLink}},
  \code{\link{Webproperty.permissions}},
  \code{\link{management.webproperties.insert}},
  \code{\link{management.webproperties.patch}},
  \code{\link{management.webproperties.update}}
}

