% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{WebPropertyRef}
\alias{WebPropertyRef}
\title{WebPropertyRef Object}
\usage{
WebPropertyRef(accountId = NULL, href = NULL, id = NULL,
  internalWebPropertyId = NULL, name = NULL)
}
\arguments{
\item{accountId}{Account ID to which this web property belongs}

\item{href}{Link for this web property}

\item{id}{Web property ID of the form UA-XXXXX-YY}

\item{internalWebPropertyId}{Internal ID for this web property}

\item{name}{Name of this web property}
}
\value{
WebPropertyRef object
}
\description{
WebPropertyRef Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for a web property reference.
}
