% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_objects.R
\name{WebPropertySummary}
\alias{WebPropertySummary}
\title{WebPropertySummary Object}
\usage{
WebPropertySummary(id = NULL, internalWebPropertyId = NULL, level = NULL,
  name = NULL, profiles = NULL, starred = NULL, websiteUrl = NULL)
}
\arguments{
\item{id}{Web property ID of the form UA-XXXXX-YY}

\item{internalWebPropertyId}{Internal ID for this web property}

\item{level}{Level for this web property}

\item{name}{Web property name}

\item{profiles}{List of profiles under this web property}

\item{starred}{Indicates whether this web property is starred or not}

\item{websiteUrl}{Website url for this web property}
}
\value{
WebPropertySummary object
}
\description{
WebPropertySummary Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for an Analytics WebPropertySummary. WebPropertySummary returns basic information (i.e., summary) for a web property.
}
