% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.remarketingAudience.get}
\alias{management.remarketingAudience.get}
\title{Gets remarketing audiences to which the user has access.}
\usage{
management.remarketingAudience.get(accountId, webPropertyId,
  remarketingAudienceId)
}
\arguments{
\item{accountId}{Account ID for the remarketing audience to retrieve}

\item{webPropertyId}{Web property ID for the remarketing audience to retrieve}

\item{remarketingAudienceId}{The ID to retrieve the Remarketing Audience for}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics.edit
\item https://www.googleapis.com/auth/analytics.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics.edit, https://www.googleapis.com/auth/analytics.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}
}

