% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.remarketingAudience.insert}
\alias{management.remarketingAudience.insert}
\title{Creates a new remarketing audience.}
\usage{
management.remarketingAudience.insert(RemarketingAudience, accountId,
  webPropertyId)
}
\arguments{
\item{RemarketingAudience}{The \link{RemarketingAudience} object to pass to this method}

\item{accountId}{The account ID for which to create the remarketing audience}

\item{webPropertyId}{Web property ID for which to create the remarketing audience}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics.edit
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics.edit)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}

Other RemarketingAudience functions: \code{\link{RemarketingAudience.audienceDefinition}},
  \code{\link{RemarketingAudience.stateBasedAudienceDefinition.excludeConditions}},
  \code{\link{RemarketingAudience.stateBasedAudienceDefinition}},
  \code{\link{RemarketingAudience}},
  \code{\link{management.remarketingAudience.patch}},
  \code{\link{management.remarketingAudience.update}}
}
